The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P11 Based on the fourth Tweedie)

The function BS13() defines the Birnbaum-Saunders distribution, a two-parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

Usage

BS13(mu.link = "log", sigma.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter (representing the scale \(\omega\)).

sigma.link

defines the sigma.link, with "log" link as the default for the sigma parameter (representing the shape \(\psi\)).

Value

Returns a gamlss.family object which can be used to fit a BS13 distribution in the gamlss() function.

Details

The Birnbaum-Saunders distribution with parameters mu and sigma (where mu represents \(\omega\) and sigma represents \(\psi\)) has density given by

\(f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{2} \left[ \frac{x\sigma}{\mu} + \frac{\mu}{x\sigma} - 2 \right] \right) \frac{[x\sigma + \mu]}{2\sqrt{\mu x^3}}\)

for \(x>0\), \(\mu>0\) and \(\sigma>0\). In this parameterization, \(E(X) = \frac{\mu}{\sigma} + \frac{\mu}{2\sigma^2}\) and \(Var(X) = \frac{\mu^2}{\sigma^3} + \frac{5\mu^2}{4\sigma^4}\).

References

Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). On new parameterizations of the Birnbaum-Saunders distribution. Pakistan Journal of Statistics, 28(1), 1-26.

See also

Author

David Villegas Ceballos, david.villegas1@udea.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
set.seed(123456)
y <- rBS13(n=500, mu=5, sigma=30)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS13,
               control=gamlss.control(n.cyc=300, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = -959.7991 
#> GAMLSS-RS iteration 2: Global Deviance = -1094.719 
#> GAMLSS-RS iteration 3: Global Deviance = -1201.023 
#> GAMLSS-RS iteration 4: Global Deviance = -1287.423 
#> GAMLSS-RS iteration 5: Global Deviance = -1359.318 
#> GAMLSS-RS iteration 6: Global Deviance = -1420.617 
#> GAMLSS-RS iteration 7: Global Deviance = -1473.785 
#> GAMLSS-RS iteration 8: Global Deviance = -1520.15 
#> GAMLSS-RS iteration 9: Global Deviance = -1562.535 
#> GAMLSS-RS iteration 10: Global Deviance = -1598.916 
#> GAMLSS-RS iteration 11: Global Deviance = -1631.619 
#> GAMLSS-RS iteration 12: Global Deviance = -1661.323 
#> GAMLSS-RS iteration 13: Global Deviance = -1688.216 
#> GAMLSS-RS iteration 14: Global Deviance = -1712.455 
#> GAMLSS-RS iteration 15: Global Deviance = -1734.883 
#> GAMLSS-RS iteration 16: Global Deviance = -1755.712 
#> GAMLSS-RS iteration 17: Global Deviance = -1774.794 
#> GAMLSS-RS iteration 18: Global Deviance = -1792.531 
#> GAMLSS-RS iteration 19: Global Deviance = -1803.669 
#> GAMLSS-RS iteration 20: Global Deviance = -1818.839 
#> GAMLSS-RS iteration 21: Global Deviance = -1833.52 
#> GAMLSS-RS iteration 22: Global Deviance = -1847.151 
#> GAMLSS-RS iteration 23: Global Deviance = -1859.903 
#> GAMLSS-RS iteration 24: Global Deviance = -1871.85 
#> GAMLSS-RS iteration 25: Global Deviance = -1882.963 
#> GAMLSS-RS iteration 26: Global Deviance = -1893.709 
#> GAMLSS-RS iteration 27: Global Deviance = -1903.667 
#> GAMLSS-RS iteration 28: Global Deviance = -1913.06 
#> GAMLSS-RS iteration 29: Global Deviance = -1921.919 
#> GAMLSS-RS iteration 30: Global Deviance = -1930.337 
#> GAMLSS-RS iteration 31: Global Deviance = -1938.295 
#> GAMLSS-RS iteration 32: Global Deviance = -1945.932 
#> GAMLSS-RS iteration 33: Global Deviance = -1952.981 
#> GAMLSS-RS iteration 34: Global Deviance = -1959.678 
#> GAMLSS-RS iteration 35: Global Deviance = -1965.949 
#> GAMLSS-RS iteration 36: Global Deviance = -1971.949 
#> GAMLSS-RS iteration 37: Global Deviance = -1977.682 
#> GAMLSS-RS iteration 38: Global Deviance = -1983.233 
#> GAMLSS-RS iteration 39: Global Deviance = -1988.593 
#> GAMLSS-RS iteration 40: Global Deviance = -1993.649 
#> GAMLSS-RS iteration 41: Global Deviance = -1998.432 
#> GAMLSS-RS iteration 42: Global Deviance = -2002.924 
#> GAMLSS-RS iteration 43: Global Deviance = -2007.219 
#> GAMLSS-RS iteration 44: Global Deviance = -2011.326 
#> GAMLSS-RS iteration 45: Global Deviance = -2015.222 
#> GAMLSS-RS iteration 46: Global Deviance = -2016.414 
#> GAMLSS-RS iteration 47: Global Deviance = -2018.35 
#> GAMLSS-RS iteration 48: Global Deviance = -2021.02 
#> GAMLSS-RS iteration 49: Global Deviance = -2024.174 
#> GAMLSS-RS iteration 50: Global Deviance = -2027.333 
#> GAMLSS-RS iteration 51: Global Deviance = -2030.589 
#> GAMLSS-RS iteration 52: Global Deviance = -2033.697 
#> GAMLSS-RS iteration 53: Global Deviance = -2036.649 
#> GAMLSS-RS iteration 54: Global Deviance = -2039.5 
#> GAMLSS-RS iteration 55: Global Deviance = -2042.222 
#> GAMLSS-RS iteration 56: Global Deviance = -2044.826 
#> GAMLSS-RS iteration 57: Global Deviance = -2047.31 
#> GAMLSS-RS iteration 58: Global Deviance = -2049.701 
#> GAMLSS-RS iteration 59: Global Deviance = -2051.992 
#> GAMLSS-RS iteration 60: Global Deviance = -2054.3 
#> GAMLSS-RS iteration 61: Global Deviance = -2056.401 
#> GAMLSS-RS iteration 62: Global Deviance = -2058.426 
#> GAMLSS-RS iteration 63: Global Deviance = -2060.461 
#> GAMLSS-RS iteration 64: Global Deviance = -2062.311 
#> GAMLSS-RS iteration 65: Global Deviance = -2064.013 
#> GAMLSS-RS iteration 66: Global Deviance = -2065.665 
#> GAMLSS-RS iteration 67: Global Deviance = -2067.33 
#> GAMLSS-RS iteration 68: Global Deviance = -2068.928 
#> GAMLSS-RS iteration 69: Global Deviance = -2070.519 
#> GAMLSS-RS iteration 70: Global Deviance = -2072.038 
#> GAMLSS-RS iteration 71: Global Deviance = -2073.489 
#> GAMLSS-RS iteration 72: Global Deviance = -2074.835 
#> GAMLSS-RS iteration 73: Global Deviance = -2076.089 
#> GAMLSS-RS iteration 74: Global Deviance = -2077.304 
#> GAMLSS-RS iteration 75: Global Deviance = -2078.48 
#> GAMLSS-RS iteration 76: Global Deviance = -2079.617 
#> GAMLSS-RS iteration 77: Global Deviance = -2080.753 
#> GAMLSS-RS iteration 78: Global Deviance = -2081.845 
#> GAMLSS-RS iteration 79: Global Deviance = -2082.931 
#> GAMLSS-RS iteration 80: Global Deviance = -2083.97 
#> GAMLSS-RS iteration 81: Global Deviance = -2084.966 
#> GAMLSS-RS iteration 82: Global Deviance = -2085.922 
#> GAMLSS-RS iteration 83: Global Deviance = -2086.84 
#> GAMLSS-RS iteration 84: Global Deviance = -2087.721 
#> GAMLSS-RS iteration 85: Global Deviance = -2088.567 
#> GAMLSS-RS iteration 86: Global Deviance = -2089.381 
#> GAMLSS-RS iteration 87: Global Deviance = -2090.143 
#> GAMLSS-RS iteration 88: Global Deviance = -2090.879 
#> GAMLSS-RS iteration 89: Global Deviance = -2091.57 
#> GAMLSS-RS iteration 90: Global Deviance = -2092.24 
#> GAMLSS-RS iteration 91: Global Deviance = -2092.888 
#> GAMLSS-RS iteration 92: Global Deviance = -2093.516 
#> GAMLSS-RS iteration 93: Global Deviance = -2094.123 
#> GAMLSS-RS iteration 94: Global Deviance = -2094.711 
#> GAMLSS-RS iteration 95: Global Deviance = -2095.28 
#> GAMLSS-RS iteration 96: Global Deviance = -2095.83 
#> GAMLSS-RS iteration 97: Global Deviance = -2096.363 
#> GAMLSS-RS iteration 98: Global Deviance = -2096.878 
#> GAMLSS-RS iteration 99: Global Deviance = -2097.376 
#> GAMLSS-RS iteration 100: Global Deviance = -2097.858 
#> GAMLSS-RS iteration 101: Global Deviance = -2098.324 
#> GAMLSS-RS iteration 102: Global Deviance = -2098.774 
#> GAMLSS-RS iteration 103: Global Deviance = -2099.21 
#> GAMLSS-RS iteration 104: Global Deviance = -2099.642 
#> GAMLSS-RS iteration 105: Global Deviance = -2100.058 
#> GAMLSS-RS iteration 106: Global Deviance = -2100.459 
#> GAMLSS-RS iteration 107: Global Deviance = -2100.856 
#> GAMLSS-RS iteration 108: Global Deviance = -2101.239 
#> GAMLSS-RS iteration 109: Global Deviance = -2101.607 
#> GAMLSS-RS iteration 110: Global Deviance = -2101.962 
#> GAMLSS-RS iteration 111: Global Deviance = -2102.304 
#> GAMLSS-RS iteration 112: Global Deviance = -2102.634 
#> GAMLSS-RS iteration 113: Global Deviance = -2102.952 
#> GAMLSS-RS iteration 114: Global Deviance = -2103.258 
#> GAMLSS-RS iteration 115: Global Deviance = -2103.554 
#> GAMLSS-RS iteration 116: Global Deviance = -2103.839 
#> GAMLSS-RS iteration 117: Global Deviance = -2104.114 
#> GAMLSS-RS iteration 118: Global Deviance = -2104.379 
#> GAMLSS-RS iteration 119: Global Deviance = -2104.635 
#> GAMLSS-RS iteration 120: Global Deviance = -2104.883 
#> GAMLSS-RS iteration 121: Global Deviance = -2105.122 
#> GAMLSS-RS iteration 122: Global Deviance = -2105.352 
#> GAMLSS-RS iteration 123: Global Deviance = -2105.575 
#> GAMLSS-RS iteration 124: Global Deviance = -2105.79 
#> GAMLSS-RS iteration 125: Global Deviance = -2105.997 
#> GAMLSS-RS iteration 126: Global Deviance = -2106.198 
#> GAMLSS-RS iteration 127: Global Deviance = -2106.392 
#> GAMLSS-RS iteration 128: Global Deviance = -2106.579 
#> GAMLSS-RS iteration 129: Global Deviance = -2106.76 
#> GAMLSS-RS iteration 130: Global Deviance = -2106.935 
#> GAMLSS-RS iteration 131: Global Deviance = -2107.104 
#> GAMLSS-RS iteration 132: Global Deviance = -2107.267 
#> GAMLSS-RS iteration 133: Global Deviance = -2107.425 
#> GAMLSS-RS iteration 134: Global Deviance = -2107.578 
#> GAMLSS-RS iteration 135: Global Deviance = -2107.725 
#> GAMLSS-RS iteration 136: Global Deviance = -2107.868 
#> GAMLSS-RS iteration 137: Global Deviance = -2108.006 
#> GAMLSS-RS iteration 138: Global Deviance = -2108.139 
#> GAMLSS-RS iteration 139: Global Deviance = -2108.268 
#> GAMLSS-RS iteration 140: Global Deviance = -2108.393 
#> GAMLSS-RS iteration 141: Global Deviance = -2108.514 
#> GAMLSS-RS iteration 142: Global Deviance = -2108.628 
#> GAMLSS-RS iteration 143: Global Deviance = -2108.739 
#> GAMLSS-RS iteration 144: Global Deviance = -2108.847 
#> GAMLSS-RS iteration 145: Global Deviance = -2108.951 
#> GAMLSS-RS iteration 146: Global Deviance = -2109.049 
#> GAMLSS-RS iteration 147: Global Deviance = -2109.145 
#> GAMLSS-RS iteration 148: Global Deviance = -2109.238 
#> GAMLSS-RS iteration 149: Global Deviance = -2109.328 
#> GAMLSS-RS iteration 150: Global Deviance = -2109.416 
#> GAMLSS-RS iteration 151: Global Deviance = -2109.5 
#> GAMLSS-RS iteration 152: Global Deviance = -2109.582 
#> GAMLSS-RS iteration 153: Global Deviance = -2109.662 
#> GAMLSS-RS iteration 154: Global Deviance = -2109.739 
#> GAMLSS-RS iteration 155: Global Deviance = -2109.814 
#> GAMLSS-RS iteration 156: Global Deviance = -2109.887 
#> GAMLSS-RS iteration 157: Global Deviance = -2109.957 
#> GAMLSS-RS iteration 158: Global Deviance = -2110.026 
#> GAMLSS-RS iteration 159: Global Deviance = -2110.092 
#> GAMLSS-RS iteration 160: Global Deviance = -2110.156 
#> GAMLSS-RS iteration 161: Global Deviance = -2110.218 
#> GAMLSS-RS iteration 162: Global Deviance = -2110.278 
#> GAMLSS-RS iteration 163: Global Deviance = -2110.337 
#> GAMLSS-RS iteration 164: Global Deviance = -2110.394 
#> GAMLSS-RS iteration 165: Global Deviance = -2110.449 
#> GAMLSS-RS iteration 166: Global Deviance = -2110.502 
#> GAMLSS-RS iteration 167: Global Deviance = -2110.554 
#> GAMLSS-RS iteration 168: Global Deviance = -2110.604 
#> GAMLSS-RS iteration 169: Global Deviance = -2110.652 
#> GAMLSS-RS iteration 170: Global Deviance = -2110.699 
#> GAMLSS-RS iteration 171: Global Deviance = -2110.745 
#> GAMLSS-RS iteration 172: Global Deviance = -2110.789 
#> GAMLSS-RS iteration 173: Global Deviance = -2110.832 
#> GAMLSS-RS iteration 174: Global Deviance = -2110.874 
#> GAMLSS-RS iteration 175: Global Deviance = -2110.914 
#> GAMLSS-RS iteration 176: Global Deviance = -2110.953 
#> GAMLSS-RS iteration 177: Global Deviance = -2110.991 
#> GAMLSS-RS iteration 178: Global Deviance = -2111.028 
#> GAMLSS-RS iteration 179: Global Deviance = -2111.063 
#> GAMLSS-RS iteration 180: Global Deviance = -2111.098 
#> GAMLSS-RS iteration 181: Global Deviance = -2111.131 
#> GAMLSS-RS iteration 182: Global Deviance = -2111.164 
#> GAMLSS-RS iteration 183: Global Deviance = -2111.195 
#> GAMLSS-RS iteration 184: Global Deviance = -2111.225 
#> GAMLSS-RS iteration 185: Global Deviance = -2111.255 
#> GAMLSS-RS iteration 186: Global Deviance = -2111.284 
#> GAMLSS-RS iteration 187: Global Deviance = -2111.311 
#> GAMLSS-RS iteration 188: Global Deviance = -2111.338 
#> GAMLSS-RS iteration 189: Global Deviance = -2111.364 
#> GAMLSS-RS iteration 190: Global Deviance = -2111.389 
#> GAMLSS-RS iteration 191: Global Deviance = -2111.414 
#> GAMLSS-RS iteration 192: Global Deviance = -2111.438 
#> GAMLSS-RS iteration 193: Global Deviance = -2111.461 
#> GAMLSS-RS iteration 194: Global Deviance = -2111.483 
#> GAMLSS-RS iteration 195: Global Deviance = -2111.505 
#> GAMLSS-RS iteration 196: Global Deviance = -2111.525 
#> GAMLSS-RS iteration 197: Global Deviance = -2111.546 
#> GAMLSS-RS iteration 198: Global Deviance = -2111.565 
#> GAMLSS-RS iteration 199: Global Deviance = -2111.584 
#> GAMLSS-RS iteration 200: Global Deviance = -2111.603 
#> GAMLSS-RS iteration 201: Global Deviance = -2111.621 
#> GAMLSS-RS iteration 202: Global Deviance = -2111.638 
#> GAMLSS-RS iteration 203: Global Deviance = -2111.655 
#> GAMLSS-RS iteration 204: Global Deviance = -2111.671 
#> GAMLSS-RS iteration 205: Global Deviance = -2111.687 
#> GAMLSS-RS iteration 206: Global Deviance = -2111.702 
#> GAMLSS-RS iteration 207: Global Deviance = -2111.717 
#> GAMLSS-RS iteration 208: Global Deviance = -2111.732 
#> GAMLSS-RS iteration 209: Global Deviance = -2111.746 
#> GAMLSS-RS iteration 210: Global Deviance = -2111.759 
#> GAMLSS-RS iteration 211: Global Deviance = -2111.772 
#> GAMLSS-RS iteration 212: Global Deviance = -2111.785 
#> GAMLSS-RS iteration 213: Global Deviance = -2111.797 
#> GAMLSS-RS iteration 214: Global Deviance = -2111.809 
#> GAMLSS-RS iteration 215: Global Deviance = -2111.821 
#> GAMLSS-RS iteration 216: Global Deviance = -2111.832 
#> GAMLSS-RS iteration 217: Global Deviance = -2111.843 
#> GAMLSS-RS iteration 218: Global Deviance = -2111.853 
#> GAMLSS-RS iteration 219: Global Deviance = -2111.864 
#> GAMLSS-RS iteration 220: Global Deviance = -2111.874 
#> GAMLSS-RS iteration 221: Global Deviance = -2111.883 
#> GAMLSS-RS iteration 222: Global Deviance = -2111.892 
#> GAMLSS-RS iteration 223: Global Deviance = -2111.901 
#> GAMLSS-RS iteration 224: Global Deviance = -2111.91 
#> GAMLSS-RS iteration 225: Global Deviance = -2111.919 
#> GAMLSS-RS iteration 226: Global Deviance = -2111.927 
#> GAMLSS-RS iteration 227: Global Deviance = -2111.935 
#> GAMLSS-RS iteration 228: Global Deviance = -2111.943 
#> GAMLSS-RS iteration 229: Global Deviance = -2111.95 
#> GAMLSS-RS iteration 230: Global Deviance = -2111.957 
#> GAMLSS-RS iteration 231: Global Deviance = -2111.964 
#> GAMLSS-RS iteration 232: Global Deviance = -2111.971 
#> GAMLSS-RS iteration 233: Global Deviance = -2111.978 
#> GAMLSS-RS iteration 234: Global Deviance = -2111.984 
#> GAMLSS-RS iteration 235: Global Deviance = -2111.99 
#> GAMLSS-RS iteration 236: Global Deviance = -2111.996 
#> GAMLSS-RS iteration 237: Global Deviance = -2112.002 
#> GAMLSS-RS iteration 238: Global Deviance = -2112.008 
#> GAMLSS-RS iteration 239: Global Deviance = -2112.013 
#> GAMLSS-RS iteration 240: Global Deviance = -2112.019 
#> GAMLSS-RS iteration 241: Global Deviance = -2112.024 
#> GAMLSS-RS iteration 242: Global Deviance = -2112.029 
#> GAMLSS-RS iteration 243: Global Deviance = -2112.034 
#> GAMLSS-RS iteration 244: Global Deviance = -2112.038 
#> GAMLSS-RS iteration 245: Global Deviance = -2112.043 
#> GAMLSS-RS iteration 246: Global Deviance = -2112.047 
#> GAMLSS-RS iteration 247: Global Deviance = -2112.052 
#> GAMLSS-RS iteration 248: Global Deviance = -2112.056 
#> GAMLSS-RS iteration 249: Global Deviance = -2112.06 
#> GAMLSS-RS iteration 250: Global Deviance = -2112.064 
#> GAMLSS-RS iteration 251: Global Deviance = -2112.067 
#> GAMLSS-RS iteration 252: Global Deviance = -2112.071 
#> GAMLSS-RS iteration 253: Global Deviance = -2112.075 
#> GAMLSS-RS iteration 254: Global Deviance = -2112.078 
#> GAMLSS-RS iteration 255: Global Deviance = -2112.081 
#> GAMLSS-RS iteration 256: Global Deviance = -2112.084 
#> GAMLSS-RS iteration 257: Global Deviance = -2112.088 
#> GAMLSS-RS iteration 258: Global Deviance = -2112.091 
#> GAMLSS-RS iteration 259: Global Deviance = -2112.094 
#> GAMLSS-RS iteration 260: Global Deviance = -2112.097 
#> GAMLSS-RS iteration 261: Global Deviance = -2112.099 
#> GAMLSS-RS iteration 262: Global Deviance = -2112.102 
#> GAMLSS-RS iteration 263: Global Deviance = -2112.104 
#> GAMLSS-RS iteration 264: Global Deviance = -2112.107 
#> GAMLSS-RS iteration 265: Global Deviance = -2112.109 
#> GAMLSS-RS iteration 266: Global Deviance = -2112.112 
#> GAMLSS-RS iteration 267: Global Deviance = -2112.114 
#> GAMLSS-RS iteration 268: Global Deviance = -2112.116 
#> GAMLSS-RS iteration 269: Global Deviance = -2112.119 
#> GAMLSS-RS iteration 270: Global Deviance = -2112.121 
#> GAMLSS-RS iteration 271: Global Deviance = -2112.123 
#> GAMLSS-RS iteration 272: Global Deviance = -2112.125 
#> GAMLSS-RS iteration 273: Global Deviance = -2112.126 
#> GAMLSS-RS iteration 274: Global Deviance = -2112.128 
#> GAMLSS-RS iteration 275: Global Deviance = -2112.13 
#> GAMLSS-RS iteration 276: Global Deviance = -2112.132 
#> GAMLSS-RS iteration 277: Global Deviance = -2112.134 
#> GAMLSS-RS iteration 278: Global Deviance = -2112.135 
#> GAMLSS-RS iteration 279: Global Deviance = -2112.137 
#> GAMLSS-RS iteration 280: Global Deviance = -2112.139 
#> GAMLSS-RS iteration 281: Global Deviance = -2112.14 
#> GAMLSS-RS iteration 282: Global Deviance = -2112.142 
#> GAMLSS-RS iteration 283: Global Deviance = -2112.143 
#> GAMLSS-RS iteration 284: Global Deviance = -2112.145 
#> GAMLSS-RS iteration 285: Global Deviance = -2112.146 
#> GAMLSS-RS iteration 286: Global Deviance = -2112.147 
#> GAMLSS-RS iteration 287: Global Deviance = -2112.148 
#> GAMLSS-RS iteration 288: Global Deviance = -2112.15 
#> GAMLSS-RS iteration 289: Global Deviance = -2112.151 
#> GAMLSS-RS iteration 290: Global Deviance = -2112.152 
#> GAMLSS-RS iteration 291: Global Deviance = -2112.153 
#> GAMLSS-RS iteration 292: Global Deviance = -2112.154 
#> GAMLSS-RS iteration 293: Global Deviance = -2112.155 
#> GAMLSS-RS iteration 294: Global Deviance = -2112.156 
#> GAMLSS-RS iteration 295: Global Deviance = -2112.157 

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
#> (Intercept) 
#>    5.239938 
exp(coef(mod1, what="sigma"))
#> (Intercept) 
#>    31.57119 

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS13
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.1 + 1.1 * x1)      # Aprox 5
  sigma <- exp(2.2 + 2.4 * x2)   # Aprox 30
  y <- rBS13(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS13, data=dat,
               control=gamlss.control(n.cyc=500, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = -354.099 
#> GAMLSS-RS iteration 2: Global Deviance = -353.136 
#> GAMLSS-RS iteration 3: Global Deviance = -451.8645 
#> GAMLSS-RS iteration 4: Global Deviance = -477.9382 
#> GAMLSS-RS iteration 5: Global Deviance = -500.3046 
#> GAMLSS-RS iteration 6: Global Deviance = -520.0801 
#> GAMLSS-RS iteration 7: Global Deviance = -537.7217 
#> GAMLSS-RS iteration 8: Global Deviance = -553.4016 
#> GAMLSS-RS iteration 9: Global Deviance = -567.5762 
#> GAMLSS-RS iteration 10: Global Deviance = -580.4127 
#> GAMLSS-RS iteration 11: Global Deviance = -592.0182 
#> GAMLSS-RS iteration 12: Global Deviance = -602.7688 
#> GAMLSS-RS iteration 13: Global Deviance = -612.8149 
#> GAMLSS-RS iteration 14: Global Deviance = -621.9787 
#> GAMLSS-RS iteration 15: Global Deviance = -630.4129 
#> GAMLSS-RS iteration 16: Global Deviance = -638.3599 
#> GAMLSS-RS iteration 17: Global Deviance = -645.7924 
#> GAMLSS-RS iteration 18: Global Deviance = -652.7076 
#> GAMLSS-RS iteration 19: Global Deviance = -659.0831 
#> GAMLSS-RS iteration 20: Global Deviance = -665.024 
#> GAMLSS-RS iteration 21: Global Deviance = -670.7233 
#> GAMLSS-RS iteration 22: Global Deviance = -676.0486 
#> GAMLSS-RS iteration 23: Global Deviance = -681.064 
#> GAMLSS-RS iteration 24: Global Deviance = -685.7759 
#> GAMLSS-RS iteration 25: Global Deviance = -690.2243 
#> GAMLSS-RS iteration 26: Global Deviance = -694.4341 
#> GAMLSS-RS iteration 27: Global Deviance = -698.4223 
#> GAMLSS-RS iteration 28: Global Deviance = -702.227 
#> GAMLSS-RS iteration 29: Global Deviance = -705.8391 
#> GAMLSS-RS iteration 30: Global Deviance = -709.1964 
#> GAMLSS-RS iteration 31: Global Deviance = -712.4328 
#> GAMLSS-RS iteration 32: Global Deviance = -715.6595 
#> GAMLSS-RS iteration 33: Global Deviance = -718.6008 
#> GAMLSS-RS iteration 34: Global Deviance = -721.3315 
#> GAMLSS-RS iteration 35: Global Deviance = -723.9652 
#> GAMLSS-RS iteration 36: Global Deviance = -726.669 
#> GAMLSS-RS iteration 37: Global Deviance = -729.218 
#> GAMLSS-RS iteration 38: Global Deviance = -731.6306 
#> GAMLSS-RS iteration 39: Global Deviance = -733.8453 
#> GAMLSS-RS iteration 40: Global Deviance = -735.9051 
#> GAMLSS-RS iteration 41: Global Deviance = -737.9004 
#> GAMLSS-RS iteration 42: Global Deviance = -739.8308 
#> GAMLSS-RS iteration 43: Global Deviance = -741.6959 
#> GAMLSS-RS iteration 44: Global Deviance = -743.4951 
#> GAMLSS-RS iteration 45: Global Deviance = -745.2874 
#> GAMLSS-RS iteration 46: Global Deviance = -747.0527 
#> GAMLSS-RS iteration 47: Global Deviance = -748.7274 
#> GAMLSS-RS iteration 48: Global Deviance = -750.3206 
#> GAMLSS-RS iteration 49: Global Deviance = -751.8398 
#> GAMLSS-RS iteration 50: Global Deviance = -753.2912 
#> GAMLSS-RS iteration 51: Global Deviance = -754.6802 
#> GAMLSS-RS iteration 52: Global Deviance = -756.0106 
#> GAMLSS-RS iteration 53: Global Deviance = -757.3087 
#> GAMLSS-RS iteration 54: Global Deviance = -758.5235 
#> GAMLSS-RS iteration 55: Global Deviance = -759.7161 
#> GAMLSS-RS iteration 56: Global Deviance = -760.8513 
#> GAMLSS-RS iteration 57: Global Deviance = -761.9621 
#> GAMLSS-RS iteration 58: Global Deviance = -762.9996 
#> GAMLSS-RS iteration 59: Global Deviance = -763.439 
#> GAMLSS-RS iteration 60: Global Deviance = -764.9138 
#> GAMLSS-RS iteration 61: Global Deviance = -766.1173 
#> GAMLSS-RS iteration 62: Global Deviance = -767.2177 
#> GAMLSS-RS iteration 63: Global Deviance = -768.0785 
#> GAMLSS-RS iteration 64: Global Deviance = -768.9105 
#> GAMLSS-RS iteration 65: Global Deviance = -769.8439 
#> GAMLSS-RS iteration 66: Global Deviance = -770.7334 
#> GAMLSS-RS iteration 67: Global Deviance = -771.4827 
#> GAMLSS-RS iteration 68: Global Deviance = -772.2876 
#> GAMLSS-RS iteration 69: Global Deviance = -772.9858 
#> GAMLSS-RS iteration 70: Global Deviance = -773.7318 
#> GAMLSS-RS iteration 71: Global Deviance = -774.3042 
#> GAMLSS-RS iteration 72: Global Deviance = -774.872 
#> GAMLSS-RS iteration 73: Global Deviance = -775.5461 
#> GAMLSS-RS iteration 74: Global Deviance = -776.1868 
#> GAMLSS-RS iteration 75: Global Deviance = -776.7948 
#> GAMLSS-RS iteration 76: Global Deviance = -777.2889 
#> GAMLSS-RS iteration 77: Global Deviance = -777.775 
#> GAMLSS-RS iteration 78: Global Deviance = -778.2946 
#> GAMLSS-RS iteration 79: Global Deviance = -778.8368 
#> GAMLSS-RS iteration 80: Global Deviance = -779.3558 
#> GAMLSS-RS iteration 81: Global Deviance = -779.8518 
#> GAMLSS-RS iteration 82: Global Deviance = -780.3262 
#> GAMLSS-RS iteration 83: Global Deviance = -780.718 
#> GAMLSS-RS iteration 84: Global Deviance = -781.1028 
#> GAMLSS-RS iteration 85: Global Deviance = -781.4793 
#> GAMLSS-RS iteration 86: Global Deviance = -781.8473 
#> GAMLSS-RS iteration 87: Global Deviance = -782.2069 
#> GAMLSS-RS iteration 88: Global Deviance = -782.5576 
#> GAMLSS-RS iteration 89: Global Deviance = -782.9254 
#> GAMLSS-RS iteration 90: Global Deviance = -783.3057 
#> GAMLSS-RS iteration 91: Global Deviance = -783.6716 
#> GAMLSS-RS iteration 92: Global Deviance = -784.0231 
#> GAMLSS-RS iteration 93: Global Deviance = -784.3609 
#> GAMLSS-RS iteration 94: Global Deviance = -784.6859 
#> GAMLSS-RS iteration 95: Global Deviance = -784.9988 
#> GAMLSS-RS iteration 96: Global Deviance = -785.3003 
#> GAMLSS-RS iteration 97: Global Deviance = -785.591 
#> GAMLSS-RS iteration 98: Global Deviance = -785.8714 
#> GAMLSS-RS iteration 99: Global Deviance = -786.142 
#> GAMLSS-RS iteration 100: Global Deviance = -786.373 
#> GAMLSS-RS iteration 101: Global Deviance = -786.5993 
#> GAMLSS-RS iteration 102: Global Deviance = -786.8202 
#> GAMLSS-RS iteration 103: Global Deviance = -787.0359 
#> GAMLSS-RS iteration 104: Global Deviance = -787.2464 
#> GAMLSS-RS iteration 105: Global Deviance = -787.4517 
#> GAMLSS-RS iteration 106: Global Deviance = -787.6521 
#> GAMLSS-RS iteration 107: Global Deviance = -787.8475 
#> GAMLSS-RS iteration 108: Global Deviance = -788.038 
#> GAMLSS-RS iteration 109: Global Deviance = -788.2237 
#> GAMLSS-RS iteration 110: Global Deviance = -788.4048 
#> GAMLSS-RS iteration 111: Global Deviance = -788.5812 
#> GAMLSS-RS iteration 112: Global Deviance = -788.7531 
#> GAMLSS-RS iteration 113: Global Deviance = -788.9206 
#> GAMLSS-RS iteration 114: Global Deviance = -789.0838 
#> GAMLSS-RS iteration 115: Global Deviance = -789.2427 
#> GAMLSS-RS iteration 116: Global Deviance = -789.3975 
#> GAMLSS-RS iteration 117: Global Deviance = -789.5482 
#> GAMLSS-RS iteration 118: Global Deviance = -789.6948 
#> GAMLSS-RS iteration 119: Global Deviance = -789.8459 
#> GAMLSS-RS iteration 120: Global Deviance = -790.0004 
#> GAMLSS-RS iteration 121: Global Deviance = -790.15 
#> GAMLSS-RS iteration 122: Global Deviance = -790.2945 
#> GAMLSS-RS iteration 123: Global Deviance = -790.4341 
#> GAMLSS-RS iteration 124: Global Deviance = -790.569 
#> GAMLSS-RS iteration 125: Global Deviance = -790.6995 
#> GAMLSS-RS iteration 126: Global Deviance = -790.8256 
#> GAMLSS-RS iteration 127: Global Deviance = -790.9477 
#> GAMLSS-RS iteration 128: Global Deviance = -791.0657 
#> GAMLSS-RS iteration 129: Global Deviance = -791.18 
#> GAMLSS-RS iteration 130: Global Deviance = -791.2906 
#> GAMLSS-RS iteration 131: Global Deviance = -791.3977 
#> GAMLSS-RS iteration 132: Global Deviance = -791.5015 
#> GAMLSS-RS iteration 133: Global Deviance = -791.6019 
#> GAMLSS-RS iteration 134: Global Deviance = -791.6993 
#> GAMLSS-RS iteration 135: Global Deviance = -791.7936 
#> GAMLSS-RS iteration 136: Global Deviance = -791.885 
#> GAMLSS-RS iteration 137: Global Deviance = -791.9736 
#> GAMLSS-RS iteration 138: Global Deviance = -792.0595 
#> GAMLSS-RS iteration 139: Global Deviance = -792.1428 
#> GAMLSS-RS iteration 140: Global Deviance = -792.2235 
#> GAMLSS-RS iteration 141: Global Deviance = -792.3019 
#> GAMLSS-RS iteration 142: Global Deviance = -792.3778 
#> GAMLSS-RS iteration 143: Global Deviance = -792.4515 
#> GAMLSS-RS iteration 144: Global Deviance = -792.5231 
#> GAMLSS-RS iteration 145: Global Deviance = -792.5925 
#> GAMLSS-RS iteration 146: Global Deviance = -792.6598 
#> GAMLSS-RS iteration 147: Global Deviance = -792.7252 
#> GAMLSS-RS iteration 148: Global Deviance = -792.7886 
#> GAMLSS-RS iteration 149: Global Deviance = -792.8502 
#> GAMLSS-RS iteration 150: Global Deviance = -792.91 
#> GAMLSS-RS iteration 151: Global Deviance = -792.9681 
#> GAMLSS-RS iteration 152: Global Deviance = -793.0245 
#> GAMLSS-RS iteration 153: Global Deviance = -793.0792 
#> GAMLSS-RS iteration 154: Global Deviance = -793.1324 
#> GAMLSS-RS iteration 155: Global Deviance = -793.1841 
#> GAMLSS-RS iteration 156: Global Deviance = -793.2342 
#> GAMLSS-RS iteration 157: Global Deviance = -793.283 
#> GAMLSS-RS iteration 158: Global Deviance = -793.3303 
#> GAMLSS-RS iteration 159: Global Deviance = -793.3738 
#> GAMLSS-RS iteration 160: Global Deviance = -793.4141 
#> GAMLSS-RS iteration 161: Global Deviance = -793.4534 
#> GAMLSS-RS iteration 162: Global Deviance = -793.4918 
#> GAMLSS-RS iteration 163: Global Deviance = -793.5293 
#> GAMLSS-RS iteration 164: Global Deviance = -793.5659 
#> GAMLSS-RS iteration 165: Global Deviance = -793.6017 
#> GAMLSS-RS iteration 166: Global Deviance = -793.6366 
#> GAMLSS-RS iteration 167: Global Deviance = -793.6707 
#> GAMLSS-RS iteration 168: Global Deviance = -793.704 
#> GAMLSS-RS iteration 169: Global Deviance = -793.7364 
#> GAMLSS-RS iteration 170: Global Deviance = -793.7681 
#> GAMLSS-RS iteration 171: Global Deviance = -793.7991 
#> GAMLSS-RS iteration 172: Global Deviance = -793.8293 
#> GAMLSS-RS iteration 173: Global Deviance = -793.8588 
#> GAMLSS-RS iteration 174: Global Deviance = -793.8876 
#> GAMLSS-RS iteration 175: Global Deviance = -793.9157 
#> GAMLSS-RS iteration 176: Global Deviance = -793.9431 
#> GAMLSS-RS iteration 177: Global Deviance = -793.9699 
#> GAMLSS-RS iteration 178: Global Deviance = -793.996 
#> GAMLSS-RS iteration 179: Global Deviance = -794.0214 
#> GAMLSS-RS iteration 180: Global Deviance = -794.0463 
#> GAMLSS-RS iteration 181: Global Deviance = -794.0706 
#> GAMLSS-RS iteration 182: Global Deviance = -794.0942 
#> GAMLSS-RS iteration 183: Global Deviance = -794.1173 
#> GAMLSS-RS iteration 184: Global Deviance = -794.1399 
#> GAMLSS-RS iteration 185: Global Deviance = -794.1619 
#> GAMLSS-RS iteration 186: Global Deviance = -794.1833 
#> GAMLSS-RS iteration 187: Global Deviance = -794.2042 
#> GAMLSS-RS iteration 188: Global Deviance = -794.2246 
#> GAMLSS-RS iteration 189: Global Deviance = -794.2446 
#> GAMLSS-RS iteration 190: Global Deviance = -794.264 
#> GAMLSS-RS iteration 191: Global Deviance = -794.2829 
#> GAMLSS-RS iteration 192: Global Deviance = -794.3014 
#> GAMLSS-RS iteration 193: Global Deviance = -794.3194 
#> GAMLSS-RS iteration 194: Global Deviance = -794.337 
#> GAMLSS-RS iteration 195: Global Deviance = -794.3542 
#> GAMLSS-RS iteration 196: Global Deviance = -794.3709 
#> GAMLSS-RS iteration 197: Global Deviance = -794.3872 
#> GAMLSS-RS iteration 198: Global Deviance = -794.4031 
#> GAMLSS-RS iteration 199: Global Deviance = -794.4187 
#> GAMLSS-RS iteration 200: Global Deviance = -794.4338 
#> GAMLSS-RS iteration 201: Global Deviance = -794.4486 
#> GAMLSS-RS iteration 202: Global Deviance = -794.463 
#> GAMLSS-RS iteration 203: Global Deviance = -794.477 
#> GAMLSS-RS iteration 204: Global Deviance = -794.4907 
#> GAMLSS-RS iteration 205: Global Deviance = -794.5041 
#> GAMLSS-RS iteration 206: Global Deviance = -794.5171 
#> GAMLSS-RS iteration 207: Global Deviance = -794.5298 
#> GAMLSS-RS iteration 208: Global Deviance = -794.5422 
#> GAMLSS-RS iteration 209: Global Deviance = -794.5542 
#> GAMLSS-RS iteration 210: Global Deviance = -794.566 
#> GAMLSS-RS iteration 211: Global Deviance = -794.5775 
#> GAMLSS-RS iteration 212: Global Deviance = -794.5887 
#> GAMLSS-RS iteration 213: Global Deviance = -794.5996 
#> GAMLSS-RS iteration 214: Global Deviance = -794.6103 
#> GAMLSS-RS iteration 215: Global Deviance = -794.6206 
#> GAMLSS-RS iteration 216: Global Deviance = -794.6308 
#> GAMLSS-RS iteration 217: Global Deviance = -794.6406 
#> GAMLSS-RS iteration 218: Global Deviance = -794.6502 
#> GAMLSS-RS iteration 219: Global Deviance = -794.6596 
#> GAMLSS-RS iteration 220: Global Deviance = -794.6688 
#> GAMLSS-RS iteration 221: Global Deviance = -794.6777 
#> GAMLSS-RS iteration 222: Global Deviance = -794.6864 
#> GAMLSS-RS iteration 223: Global Deviance = -794.6949 
#> GAMLSS-RS iteration 224: Global Deviance = -794.7031 
#> GAMLSS-RS iteration 225: Global Deviance = -794.7112 
#> GAMLSS-RS iteration 226: Global Deviance = -794.719 
#> GAMLSS-RS iteration 227: Global Deviance = -794.7267 
#> GAMLSS-RS iteration 228: Global Deviance = -794.7342 
#> GAMLSS-RS iteration 229: Global Deviance = -794.7414 
#> GAMLSS-RS iteration 230: Global Deviance = -794.7485 
#> GAMLSS-RS iteration 231: Global Deviance = -794.7554 
#> GAMLSS-RS iteration 232: Global Deviance = -794.7622 
#> GAMLSS-RS iteration 233: Global Deviance = -794.7688 
#> GAMLSS-RS iteration 234: Global Deviance = -794.7752 
#> GAMLSS-RS iteration 235: Global Deviance = -794.7814 
#> GAMLSS-RS iteration 236: Global Deviance = -794.7875 
#> GAMLSS-RS iteration 237: Global Deviance = -794.7934 
#> GAMLSS-RS iteration 238: Global Deviance = -794.7992 
#> GAMLSS-RS iteration 239: Global Deviance = -794.8049 
#> GAMLSS-RS iteration 240: Global Deviance = -794.8104 
#> GAMLSS-RS iteration 241: Global Deviance = -794.8157 
#> GAMLSS-RS iteration 242: Global Deviance = -794.821 
#> GAMLSS-RS iteration 243: Global Deviance = -794.8261 
#> GAMLSS-RS iteration 244: Global Deviance = -794.831 
#> GAMLSS-RS iteration 245: Global Deviance = -794.8359 
#> GAMLSS-RS iteration 246: Global Deviance = -794.8406 
#> GAMLSS-RS iteration 247: Global Deviance = -794.8452 
#> GAMLSS-RS iteration 248: Global Deviance = -794.8497 
#> GAMLSS-RS iteration 249: Global Deviance = -794.8541 
#> GAMLSS-RS iteration 250: Global Deviance = -794.8583 
#> GAMLSS-RS iteration 251: Global Deviance = -794.8625 
#> GAMLSS-RS iteration 252: Global Deviance = -794.8665 
#> GAMLSS-RS iteration 253: Global Deviance = -794.8705 
#> GAMLSS-RS iteration 254: Global Deviance = -794.8743 
#> GAMLSS-RS iteration 255: Global Deviance = -794.8781 
#> GAMLSS-RS iteration 256: Global Deviance = -794.8817 
#> GAMLSS-RS iteration 257: Global Deviance = -794.8853 
#> GAMLSS-RS iteration 258: Global Deviance = -794.8888 
#> GAMLSS-RS iteration 259: Global Deviance = -794.8922 
#> GAMLSS-RS iteration 260: Global Deviance = -794.8955 
#> GAMLSS-RS iteration 261: Global Deviance = -794.8987 
#> GAMLSS-RS iteration 262: Global Deviance = -794.9018 
#> GAMLSS-RS iteration 263: Global Deviance = -794.9049 
#> GAMLSS-RS iteration 264: Global Deviance = -794.9079 
#> GAMLSS-RS iteration 265: Global Deviance = -794.9108 
#> GAMLSS-RS iteration 266: Global Deviance = -794.9136 
#> GAMLSS-RS iteration 267: Global Deviance = -794.9163 
#> GAMLSS-RS iteration 268: Global Deviance = -794.919 
#> GAMLSS-RS iteration 269: Global Deviance = -794.9217 
#> GAMLSS-RS iteration 270: Global Deviance = -794.9242 
#> GAMLSS-RS iteration 271: Global Deviance = -794.9267 
#> GAMLSS-RS iteration 272: Global Deviance = -794.9291 
#> GAMLSS-RS iteration 273: Global Deviance = -794.9315 
#> GAMLSS-RS iteration 274: Global Deviance = -794.9338 
#> GAMLSS-RS iteration 275: Global Deviance = -794.9361 
#> GAMLSS-RS iteration 276: Global Deviance = -794.9383 
#> GAMLSS-RS iteration 277: Global Deviance = -794.9404 
#> GAMLSS-RS iteration 278: Global Deviance = -794.9425 
#> GAMLSS-RS iteration 279: Global Deviance = -794.9445 
#> GAMLSS-RS iteration 280: Global Deviance = -794.9467 
#> GAMLSS-RS iteration 281: Global Deviance = -794.9488 
#> GAMLSS-RS iteration 282: Global Deviance = -794.9508 
#> GAMLSS-RS iteration 283: Global Deviance = -794.9528 
#> GAMLSS-RS iteration 284: Global Deviance = -794.9547 
#> GAMLSS-RS iteration 285: Global Deviance = -794.9566 
#> GAMLSS-RS iteration 286: Global Deviance = -794.9586 
#> GAMLSS-RS iteration 287: Global Deviance = -794.9605 
#> GAMLSS-RS iteration 288: Global Deviance = -794.9624 
#> GAMLSS-RS iteration 289: Global Deviance = -794.9642 
#> GAMLSS-RS iteration 290: Global Deviance = -794.966 
#> GAMLSS-RS iteration 291: Global Deviance = -794.9677 
#> GAMLSS-RS iteration 292: Global Deviance = -794.9694 
#> GAMLSS-RS iteration 293: Global Deviance = -794.971 
#> GAMLSS-RS iteration 294: Global Deviance = -794.9725 
#> GAMLSS-RS iteration 295: Global Deviance = -794.974 
#> GAMLSS-RS iteration 296: Global Deviance = -794.9755 
#> GAMLSS-RS iteration 297: Global Deviance = -794.9769 
#> GAMLSS-RS iteration 298: Global Deviance = -794.9783 
#> GAMLSS-RS iteration 299: Global Deviance = -794.9796 
#> GAMLSS-RS iteration 300: Global Deviance = -794.9809 
#> GAMLSS-RS iteration 301: Global Deviance = -794.9821 
#> GAMLSS-RS iteration 302: Global Deviance = -794.9834 
#> GAMLSS-RS iteration 303: Global Deviance = -794.9845 
#> GAMLSS-RS iteration 304: Global Deviance = -794.9857 
#> GAMLSS-RS iteration 305: Global Deviance = -794.9868 
#> GAMLSS-RS iteration 306: Global Deviance = -794.9878 
#> GAMLSS-RS iteration 307: Global Deviance = -794.9889 
#> GAMLSS-RS iteration 308: Global Deviance = -794.9899 
#> GAMLSS-RS iteration 309: Global Deviance = -794.9909 

summary(mod2)
#> Warning: summary: vcov has failed, option qr is used instead
#> ******************************************************************
#> Family:  c("BS13", "Birnbaum-Saunders - Thirteenth parameterization") 
#> 
#> Call:  gamlss(formula = y ~ x1, sigma.formula = ~x2, family = BS13,  
#>     data = dat, control = gamlss.control(n.cyc = 500, trace = TRUE)) 
#> 
#> Fitting method: RS() 
#> 
#> ------------------------------------------------------------------
#> Mu link function:  log
#> Mu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.12047    0.02428   46.16   <2e-16 ***
#> x1           1.05391    0.04389   24.01   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ------------------------------------------------------------------
#> Sigma link function:  log
#> Sigma Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.15332    0.03189   67.53   <2e-16 ***
#> x2           2.47011    0.04375   56.45   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ------------------------------------------------------------------
#> No. of observations in the fit:  200 
#> Degrees of Freedom for the fit:  4
#>       Residual Deg. of Freedom:  196 
#>                       at cycle:  309 
#>  
#> Global Deviance:     -794.9909 
#>             AIC:     -786.9909 
#>             SBC:     -773.7976 
#> ******************************************************************