The Beta Weibull family

The Beta Weibull family

BW(mu.link = "log", sigma.link = "log", nu.link = "log",
  tau.link = "log")

Arguments

mu.link
defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link
defines the sigma.link, with "log" link as the default for the sigma.

nu.link
defines the nu.link, with "log" link as the default for the nu parameter.

tau.link
defines the tau.link, with "log" link as the default for the tau parameter.

Details

The Beta-Weibull distribution with parameters mu, sigma, nu and tau has density given by

\(f(x)= \frac{1}{B(\nu, \tau)} \mu \sigma x^{\sigma - 1} (1 - \exp(-\mu x^\sigma))^{\nu - 1} \exp(-\mu \tau x^\sigma),\)

for \(x > 0\), \(\mu > 0\), \(\sigma > 0\), \(\nu > 0\) and \(\tau > 0\).

References

Almalki SJ, Nadarajah S (2014). “Modifications of the Weibull distribution: A review.” Reliability Engineering \& System Safety, 124, 32--55.

Famoye F, Lee C, Olumolade O (2005). “The beta-Weibull distribution.” Journal of Statistical Theory and Applications, 4(2), 121--136.

See also

Examples

# Example 1
# Generating some random values with
# known mu, sigma, nu and tau
y <- rBW(n=100, mu = (1/4), sigma =1, nu=1, tau=2)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, tau.fo=~1, family='BW',
             control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma, nu and tau
# using the inverse link function
exp(coef(mod, what='mu'))

#> (Intercept) 
#>   0.4438045 

exp(coef(mod, what='sigma'))

#> (Intercept) 
#>    1.027126 

exp(coef(mod, what='nu'))

#> (Intercept) 
#>    1.009956 

exp(coef(mod, what='tau'))

#> (Intercept) 
#>     1.12605 

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(0.75 - x1)
sigma <- exp(0.5 - x2)
nu <- 1
tau <- 2
x <- rBW(n=n, mu, sigma, nu, tau)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, tau.fo=~1, family=BW,
             control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")

#> (Intercept)          x1 
#>    2.987531   -3.443720 

coef(mod, what="sigma")

#> (Intercept)          x2 
#>   0.1830297  -0.3518636 

exp(coef(mod, what="nu"))

#> (Intercept) 
#>   0.9699999 

exp(coef(mod, what="tau"))

#> (Intercept) 
#>   0.8238145