The Extended Odd Frechet-Nadarajah-Haghighi family

The Extended Odd Frechet-Nadarjad-Hanhighi family

Usage

EOFNH(mu.link = "log", sigma.link = "log", nu.link = "log", tau.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

nu.link

defines the nu.link, with "log" link as the default for the nu parameter.

tau.link

defines the tau.link, with "log" link as the default for the tau parameter.

Value

Returns a gamlss.family object which can be used to fit a EOFNH distribution in the gamlss() function.

Details

The Extended Odd Frechet-Nadarajah-Haghighi distribution with parameters mu, sigma, nu and tau has density given by

\(f(x)= \frac{\mu\sigma\nu\tau(1+\nu x)^{\sigma-1}e^{(1-(1+\nu x)^\sigma)}[1-(1-e^{(1-(1+\nu x)^\sigma)})^{\mu}]^{\tau-1}}{(1-e^{(1-(1+\nu x)^{\sigma})})^{\mu\tau+1}} e^{-[(1-e^{(1-(1+\nu x)^\sigma)})^{-\mu}-1]^{\tau}},\)

for \(x > 0\), \(\mu > 0\), \(\sigma > 0\), \(\nu > 0\) and \(\tau > 0\).

References

Nasiru, S. (2018). Extended Odd Fréchet‐G Family of Distributions Journal of Probability and Statistics, 2018(1), 2931326.

nasiru2018extendedRelDists

See also

Author

Helber Santiago Padilla, hspadillar@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma, nu and tau
set.seed(123)
y <- rEOFNH(n=100, mu=1, sigma=2.1, nu=0.8, tau=1)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, tau.fo=~1, family=EOFNH,
              control=gamlss.control(n.cyc=5000, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 17.9417 
#> GAMLSS-RS iteration 2: Global Deviance = 15.3461 
#> GAMLSS-RS iteration 3: Global Deviance = 15.2991 
#> GAMLSS-RS iteration 4: Global Deviance = 15.252 
#> GAMLSS-RS iteration 5: Global Deviance = 15.2055 
#> GAMLSS-RS iteration 6: Global Deviance = 15.1597 
#> GAMLSS-RS iteration 7: Global Deviance = 15.1147 
#> GAMLSS-RS iteration 8: Global Deviance = 15.0702 
#> GAMLSS-RS iteration 9: Global Deviance = 15.0265 
#> GAMLSS-RS iteration 10: Global Deviance = 14.9834 
#> GAMLSS-RS iteration 11: Global Deviance = 14.9409 
#> GAMLSS-RS iteration 12: Global Deviance = 14.8992 
#> GAMLSS-RS iteration 13: Global Deviance = 14.8581 
#> GAMLSS-RS iteration 14: Global Deviance = 14.8176 
#> GAMLSS-RS iteration 15: Global Deviance = 14.7779 
#> GAMLSS-RS iteration 16: Global Deviance = 14.7388 
#> GAMLSS-RS iteration 17: Global Deviance = 14.7003 
#> GAMLSS-RS iteration 18: Global Deviance = 14.6626 
#> GAMLSS-RS iteration 19: Global Deviance = 14.6254 
#> GAMLSS-RS iteration 20: Global Deviance = 14.5889 
#> GAMLSS-RS iteration 21: Global Deviance = 14.5531 
#> GAMLSS-RS iteration 22: Global Deviance = 14.5179 
#> GAMLSS-RS iteration 23: Global Deviance = 14.4834 
#> GAMLSS-RS iteration 24: Global Deviance = 14.4494 
#> GAMLSS-RS iteration 25: Global Deviance = 14.4161 
#> GAMLSS-RS iteration 26: Global Deviance = 14.3835 
#> GAMLSS-RS iteration 27: Global Deviance = 14.3514 
#> GAMLSS-RS iteration 28: Global Deviance = 14.32 
#> GAMLSS-RS iteration 29: Global Deviance = 14.2892 
#> GAMLSS-RS iteration 30: Global Deviance = 14.259 
#> GAMLSS-RS iteration 31: Global Deviance = 14.2293 
#> GAMLSS-RS iteration 32: Global Deviance = 14.2003 
#> GAMLSS-RS iteration 33: Global Deviance = 14.1719 
#> GAMLSS-RS iteration 34: Global Deviance = 14.144 
#> GAMLSS-RS iteration 35: Global Deviance = 14.1168 
#> GAMLSS-RS iteration 36: Global Deviance = 14.0901 
#> GAMLSS-RS iteration 37: Global Deviance = 14.0639 
#> GAMLSS-RS iteration 38: Global Deviance = 14.0384 
#> GAMLSS-RS iteration 39: Global Deviance = 14.0134 
#> GAMLSS-RS iteration 40: Global Deviance = 13.9889 
#> GAMLSS-RS iteration 41: Global Deviance = 13.965 
#> GAMLSS-RS iteration 42: Global Deviance = 13.9416 
#> GAMLSS-RS iteration 43: Global Deviance = 13.9187 
#> GAMLSS-RS iteration 44: Global Deviance = 13.8964 
#> GAMLSS-RS iteration 45: Global Deviance = 13.8745 
#> GAMLSS-RS iteration 46: Global Deviance = 13.8532 
#> GAMLSS-RS iteration 47: Global Deviance = 13.8324 
#> GAMLSS-RS iteration 48: Global Deviance = 13.8121 
#> GAMLSS-RS iteration 49: Global Deviance = 13.7922 
#> GAMLSS-RS iteration 50: Global Deviance = 13.7729 
#> GAMLSS-RS iteration 51: Global Deviance = 13.754 
#> GAMLSS-RS iteration 52: Global Deviance = 13.7355 
#> GAMLSS-RS iteration 53: Global Deviance = 13.7176 
#> GAMLSS-RS iteration 54: Global Deviance = 13.7 
#> GAMLSS-RS iteration 55: Global Deviance = 13.683 
#> GAMLSS-RS iteration 56: Global Deviance = 13.6663 
#> GAMLSS-RS iteration 57: Global Deviance = 13.6501 
#> GAMLSS-RS iteration 58: Global Deviance = 13.6343 
#> GAMLSS-RS iteration 59: Global Deviance = 13.6189 
#> GAMLSS-RS iteration 60: Global Deviance = 13.604 
#> GAMLSS-RS iteration 61: Global Deviance = 13.5894 
#> GAMLSS-RS iteration 62: Global Deviance = 13.5752 
#> GAMLSS-RS iteration 63: Global Deviance = 13.5614 
#> GAMLSS-RS iteration 64: Global Deviance = 13.5479 
#> GAMLSS-RS iteration 65: Global Deviance = 13.5348 
#> GAMLSS-RS iteration 66: Global Deviance = 13.5221 
#> GAMLSS-RS iteration 67: Global Deviance = 13.5098 
#> GAMLSS-RS iteration 68: Global Deviance = 13.4978 
#> GAMLSS-RS iteration 69: Global Deviance = 13.4867 
#> GAMLSS-RS iteration 70: Global Deviance = 13.4755 
#> GAMLSS-RS iteration 71: Global Deviance = 13.465 
#> GAMLSS-RS iteration 72: Global Deviance = 13.4543 
#> GAMLSS-RS iteration 73: Global Deviance = 13.4444 
#> GAMLSS-RS iteration 74: Global Deviance = 13.4343 
#> GAMLSS-RS iteration 75: Global Deviance = 13.425 
#> GAMLSS-RS iteration 76: Global Deviance = 13.4155 
#> GAMLSS-RS iteration 77: Global Deviance = 13.4067 
#> GAMLSS-RS iteration 78: Global Deviance = 13.3991 
#> GAMLSS-RS iteration 79: Global Deviance = 13.3908 
#> GAMLSS-RS iteration 80: Global Deviance = 13.3835 
#> GAMLSS-RS iteration 81: Global Deviance = 13.3764 
#> GAMLSS-RS iteration 82: Global Deviance = 13.3695 
#> GAMLSS-RS iteration 83: Global Deviance = 13.3627 
#> GAMLSS-RS iteration 84: Global Deviance = 13.3561 
#> GAMLSS-RS iteration 85: Global Deviance = 13.3497 
#> GAMLSS-RS iteration 86: Global Deviance = 13.3434 
#> GAMLSS-RS iteration 87: Global Deviance = 13.3373 
#> GAMLSS-RS iteration 88: Global Deviance = 13.3313 
#> GAMLSS-RS iteration 89: Global Deviance = 13.3255 
#> GAMLSS-RS iteration 90: Global Deviance = 13.3199 
#> GAMLSS-RS iteration 91: Global Deviance = 13.3143 
#> GAMLSS-RS iteration 92: Global Deviance = 13.309 
#> GAMLSS-RS iteration 93: Global Deviance = 13.3037 
#> GAMLSS-RS iteration 94: Global Deviance = 13.2986 
#> GAMLSS-RS iteration 95: Global Deviance = 13.2936 
#> GAMLSS-RS iteration 96: Global Deviance = 13.2888 
#> GAMLSS-RS iteration 97: Global Deviance = 13.2841 
#> GAMLSS-RS iteration 98: Global Deviance = 13.2795 
#> GAMLSS-RS iteration 99: Global Deviance = 13.275 
#> GAMLSS-RS iteration 100: Global Deviance = 13.2706 
#> GAMLSS-RS iteration 101: Global Deviance = 13.2664 
#> GAMLSS-RS iteration 102: Global Deviance = 13.2622 
#> GAMLSS-RS iteration 103: Global Deviance = 13.2582 
#> GAMLSS-RS iteration 104: Global Deviance = 13.2543 
#> GAMLSS-RS iteration 105: Global Deviance = 13.2505 
#> GAMLSS-RS iteration 106: Global Deviance = 13.2467 
#> GAMLSS-RS iteration 107: Global Deviance = 13.2431 
#> GAMLSS-RS iteration 108: Global Deviance = 13.2396 
#> GAMLSS-RS iteration 109: Global Deviance = 13.2361 
#> GAMLSS-RS iteration 110: Global Deviance = 13.2328 
#> GAMLSS-RS iteration 111: Global Deviance = 13.2295 
#> GAMLSS-RS iteration 112: Global Deviance = 13.2264 
#> GAMLSS-RS iteration 113: Global Deviance = 13.2233 
#> GAMLSS-RS iteration 114: Global Deviance = 13.2203 
#> GAMLSS-RS iteration 115: Global Deviance = 13.2174 
#> GAMLSS-RS iteration 116: Global Deviance = 13.2145 
#> GAMLSS-RS iteration 117: Global Deviance = 13.2117 
#> GAMLSS-RS iteration 118: Global Deviance = 13.209 
#> GAMLSS-RS iteration 119: Global Deviance = 13.2064 
#> GAMLSS-RS iteration 120: Global Deviance = 13.2039 
#> GAMLSS-RS iteration 121: Global Deviance = 13.2014 
#> GAMLSS-RS iteration 122: Global Deviance = 13.1989 
#> GAMLSS-RS iteration 123: Global Deviance = 13.1966 
#> GAMLSS-RS iteration 124: Global Deviance = 13.1943 
#> GAMLSS-RS iteration 125: Global Deviance = 13.1921 
#> GAMLSS-RS iteration 126: Global Deviance = 13.1899 
#> GAMLSS-RS iteration 127: Global Deviance = 13.1878 
#> GAMLSS-RS iteration 128: Global Deviance = 13.1857 
#> GAMLSS-RS iteration 129: Global Deviance = 13.1837 
#> GAMLSS-RS iteration 130: Global Deviance = 13.1817 
#> GAMLSS-RS iteration 131: Global Deviance = 13.1798 
#> GAMLSS-RS iteration 132: Global Deviance = 13.178 
#> GAMLSS-RS iteration 133: Global Deviance = 13.1762 
#> GAMLSS-RS iteration 134: Global Deviance = 13.1744 
#> GAMLSS-RS iteration 135: Global Deviance = 13.1727 
#> GAMLSS-RS iteration 136: Global Deviance = 13.171 
#> GAMLSS-RS iteration 137: Global Deviance = 13.1693 
#> GAMLSS-RS iteration 138: Global Deviance = 13.1677 
#> GAMLSS-RS iteration 139: Global Deviance = 13.1662 
#> GAMLSS-RS iteration 140: Global Deviance = 13.1646 
#> GAMLSS-RS iteration 141: Global Deviance = 13.1632 
#> GAMLSS-RS iteration 142: Global Deviance = 13.1617 
#> GAMLSS-RS iteration 143: Global Deviance = 13.1603 
#> GAMLSS-RS iteration 144: Global Deviance = 13.159 
#> GAMLSS-RS iteration 145: Global Deviance = 13.1576 
#> GAMLSS-RS iteration 146: Global Deviance = 13.1563 
#> GAMLSS-RS iteration 147: Global Deviance = 13.1551 
#> GAMLSS-RS iteration 148: Global Deviance = 13.1538 
#> GAMLSS-RS iteration 149: Global Deviance = 13.1527 
#> GAMLSS-RS iteration 150: Global Deviance = 13.1515 
#> GAMLSS-RS iteration 151: Global Deviance = 13.1504 
#> GAMLSS-RS iteration 152: Global Deviance = 13.1493 
#> GAMLSS-RS iteration 153: Global Deviance = 13.1482 
#> GAMLSS-RS iteration 154: Global Deviance = 13.1471 
#> GAMLSS-RS iteration 155: Global Deviance = 13.1461 
#> GAMLSS-RS iteration 156: Global Deviance = 13.1451 

# Extracting the fitted values for mu, sigma, nu and tau
# using the inverse link function
exp(coef(mod, what="mu"))
#> (Intercept) 
#>   0.9036802 
exp(coef(mod, what="sigma"))
#> (Intercept) 
#>     2.55407 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>   0.5872532 
exp(coef(mod, what="tau"))
#> (Intercept) 
#>    1.075305 

# Example 2
# Generating random values under the model
n <- 100
x1 <- runif(n)
x2 <- runif(n)
mu <- exp(0.5 - 1.2 * x1)
sigma <- 2.1
nu <- 0.8
tau <- 1
y <- rEOFNH(n=n, mu, sigma, nu, tau)

mod <- gamlss(y~x1, sigma.fo=~1, nu.fo=~1, tau.fo=~1, family=EOFNH,
              control=gamlss.control(n.cyc=5000, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 17.767 
#> GAMLSS-RS iteration 2: Global Deviance = 15.6345 
#> GAMLSS-RS iteration 3: Global Deviance = 15.6326 
#> GAMLSS-RS iteration 4: Global Deviance = 15.6322 

coef(mod, what="mu")
#> (Intercept)          x1 
#>  0.05375426 -0.99943981 
exp(coef(mod, what="sigma"))
#> (Intercept) 
#>    1.330352 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>   0.9497364 
exp(coef(mod, what="tau"))
#> (Intercept) 
#>    1.382512