The Exponentiated Weibull family

The Exponentiated Weibull family

EW(mu.link = "log", sigma.link = "log", nu.link = "log")

Arguments

mu.link
defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link
defines the sigma.link, with "log" link as the default for the sigma.

nu.link
defines the nu.link, with "log" link as the default for the nu parameter.

Details

The Exponentiated Weibull Distribution with parameters mu, sigma and nu has density given by

\(f(x)=\nu \mu \sigma x^{\sigma-1} \exp(-\mu x^\sigma) (1-\exp(-\mu x^\sigma))^{\nu-1},\)

for x > 0.

See also

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rEW(n=100, mu=2, sigma=1.5, nu=0.5)

# Fitting the model
require(gamlss)

#> Loading required package: gamlss

#> Loading required package: splines

#> Loading required package: gamlss.data

#> Loading required package: gamlss.dist

#> Loading required package: MASS

#> Loading required package: nlme

#> Loading required package: parallel

#>  **********   GAMLSS Version 5.1-2  ********** 

#> For more on GAMLSS look at http://www.gamlss.org/

#> Type gamlssNews() to see new features/changes/bug fixes.

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family='EW',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma and nu
# using the inverse link function
exp(coef(mod, what='mu'))

#> (Intercept) 
#>       2.566 

exp(coef(mod, what='sigma'))

#> (Intercept) 
#>    1.155836 

exp(coef(mod, what='nu'))

#> (Intercept) 
#>   0.7234342 

# Example 2
# Generating random values under some model
n <- 200
x1 <- rpois(n, lambda=2)
x2 <- runif(n)
mu <- exp(2 + -3 * x1)
sigma <- exp(3 - 2 * x2)
nu <- 2
x <- rEW(n=n, mu, sigma, nu)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, family=EW,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")

#> (Intercept)          x1 
#>    1.955719   -2.942370 

coef(mod, what="sigma")

#> (Intercept)          x2 
#>    2.968294   -1.987047 

coef(mod, what="nu")

#> (Intercept) 
#>   0.7016593