The exponentiated XLindley family

The function EXL() defines The exponentiated XLindley, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

Usage

EXL(mu.link = "log", sigma.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

Value

Returns a gamlss.family object which can be used to fit a EXL distribution in the gamlss() function.

Details

The exponentiated XLindley with parameters mu and sigma has density given by

$$f(x) = \frac{\sigma\mu^2(2+\mu + x)\exp(-\mu x)}{(1+\mu)^2}\left[1- \left(1+\frac{\mu x}{(1 + \mu)^2}\right) \exp(-\mu x)\right] ^ {\sigma-1} $$ for \(x \geq 0\), \(\mu \geq 0\) and \(\sigma \geq 0\).

Note: In this implementation we changed the original parameters \(\delta\) for \(\mu\) and \(\alpha\) for \(\sigma\) we did it to implement this distribution within gamlss framework.

References

Alomair, A. M., Ahmed, M., Tariq, S., Ahsan-ul-Haq, M., & Talib, J. (2024). An exponentiated XLindley distribution with properties, inference and applications. Heliyon, 10(3).

See also

Author

Manuel Gutierrez Tangarife, mgutierrezta@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
y <- rEXL(n=1000, mu=0.75, sigma=1.3)

# Fitting the model
require(gamlss)
mod <- gamlss(y~1, sigma.fo=~1, family="EXL")
#> GAMLSS-RS iteration 1: Global Deviance = 3417.025 
#> GAMLSS-RS iteration 2: Global Deviance = 3401.742 
#> GAMLSS-RS iteration 3: Global Deviance = 3397.494 
#> GAMLSS-RS iteration 4: Global Deviance = 3396.325 
#> GAMLSS-RS iteration 5: Global Deviance = 3396.006 
#> GAMLSS-RS iteration 6: Global Deviance = 3395.919 
#> GAMLSS-RS iteration 7: Global Deviance = 3395.896 
#> GAMLSS-RS iteration 8: Global Deviance = 3395.889 
#> GAMLSS-RS iteration 9: Global Deviance = 3395.887 
#> GAMLSS-RS iteration 10: Global Deviance = 3395.887 

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod, what="mu"))
#> (Intercept) 
#>   0.7590244 
exp(coef(mod, what="sigma"))
#> (Intercept) 
#>    1.347231 

# Example 2
# Generating random values under some model
n <- 1000
x1 <- runif(n)
x2 <- runif(n)
mu <- exp(1.45 - 3 * x1)
sigma <- exp(2 - 1.5 * x2)
y <- rEXL(n=n, mu=mu, sigma=sigma)

mod <- gamlss(y~x1, sigma.fo=~x2, family=EXL)
#> GAMLSS-RS iteration 1: Global Deviance = 3671.753 
#> GAMLSS-RS iteration 2: Global Deviance = 3521.592 
#> GAMLSS-RS iteration 3: Global Deviance = 3452.472 
#> GAMLSS-RS iteration 4: Global Deviance = 3420.754 
#> GAMLSS-RS iteration 5: Global Deviance = 3405.964 
#> GAMLSS-RS iteration 6: Global Deviance = 3399.055 
#> GAMLSS-RS iteration 7: Global Deviance = 3395.819 
#> GAMLSS-RS iteration 8: Global Deviance = 3394.29 
#> GAMLSS-RS iteration 9: Global Deviance = 3393.576 
#> GAMLSS-RS iteration 10: Global Deviance = 3393.242 
#> GAMLSS-RS iteration 11: Global Deviance = 3393.086 
#> GAMLSS-RS iteration 12: Global Deviance = 3393.012 
#> GAMLSS-RS iteration 13: Global Deviance = 3392.977 
#> GAMLSS-RS iteration 14: Global Deviance = 3392.961 
#> GAMLSS-RS iteration 15: Global Deviance = 3392.953 
#> GAMLSS-RS iteration 16: Global Deviance = 3392.949 
#> GAMLSS-RS iteration 17: Global Deviance = 3392.947 
#> GAMLSS-RS iteration 18: Global Deviance = 3392.947 

summary(mod)
#> ******************************************************************
#> Family:  c("EXL", "Exponentiated XLindley") 
#> 
#> Call:  gamlss(formula = y ~ x1, sigma.formula = ~x2, family = EXL) 
#> 
#> Fitting method: RS() 
#> 
#> ------------------------------------------------------------------
#> Mu link function:  log
#> Mu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.43774    0.04262   33.74   <2e-16 ***
#> x1          -3.03625    0.05571  -54.50   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ------------------------------------------------------------------
#> Sigma link function:  log
#> Sigma Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.87507    0.08935   20.98   <2e-16 ***
#> x2          -1.33658    0.11731  -11.39   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ------------------------------------------------------------------
#> No. of observations in the fit:  1000 
#> Degrees of Freedom for the fit:  4
#>       Residual Deg. of Freedom:  996 
#>                       at cycle:  18 
#>  
#> Global Deviance:     3392.947 
#>             AIC:     3400.947 
#>             SBC:     3420.578 
#> ******************************************************************