The Four-Parameter Exponentiated Generalized Gamma family

The Four-Parameter Exponentiated Generalized Gamma distribution

FPEGG(mu.link = "log", sigma.link = "log", nu.link = "log", tau.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

nu.link

defines the nu.link, with "log" link as the default for the nu parameter.

tau.link

defines the tau.link, with "log" link as the default for the tau parameter.

Details

Four-Parameter Exponentiated Generalized Gamma distribution with parameters mu, sigma, nu and tau has density given by

\(f(x) = \frac {\nu \sigma}{\mu \tau(\tau)} (\frac {x}{\mu})^{\sigma \tau -1} exp\{(-{\frac {x}{\mu})^{\sigma}})\} \{ \gamma~_1~ (\tau, (\frac {x}{\mu})^{\sigma}) \}^{\nu -1} ,\)

for x > 0.

References

Almalki SJ, Nadarajah S (2014). “Modifications of the Weibull distribution: A review.” Reliability Engineering & System Safety, 124, 32--55.

Gauss M. C, Edwin M.M O, Giovana O. S (2011). “The exponentiated generalized gamma distribution with application to lifetime data.” Journal of Statistical Computation and Simulation, 81(7), 827--842. doi:10.1080/00949650903517874 , https://www.tandfonline.com/doi/abs/10.1080/00949650903517874.

See also

Author

Amylkar Urrea Montoya, amylkar.urrea@udea.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma, nu and tau
y <- rFPEGG(n=200, mu=0.1, sigma=0.8, nu=10, tau=1.5)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, tau.fo=~1, family='FPEGG',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma, nu and tau
# using the inverse link function
exp(coef(mod, what='mu'))
#> (Intercept) 
#>    0.346169 
exp(coef(mod, what='sigma'))
#> (Intercept) 
#>    1.297827 
exp(coef(mod, what='nu'))
#> (Intercept) 
#>    7.664145 
exp(coef(mod, what='tau'))
#> (Intercept) 
#>   0.5598359 

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(-0.8 + -3 * x1)
sigma <- exp(0.77 - 2 * x2)
nu <- 10
tau <- 1.5
x <- rFPEGG(n=n, mu, sigma, nu, tau)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, tau.fo=~1, family=FPEGG,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
#> (Intercept)          x1 
#>  -0.9799606  -2.4498325 
coef(mod, what="sigma")
#> (Intercept)          x2 
#>    1.018480   -2.592002 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>    30.98329 
exp(coef(mod, what="tau"))
#> (Intercept) 
#>   0.7211817