The Lindley family

The function LIN() defines the Lindley distribution with only one parameter for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

LIN(mu.link = "log")

Arguments

mu.link
defines the mu.link, with "log" link as the default for the mu parameter.

Details

The Lindley with parameter mu has density given by

\(f(x) = \frac{\mu^2}{\mu+1} (1+x) \exp(-\mu x),\)

for x > 0.

References

Lindley DV (1958). “Fiducial distributions and Bayes' theorem.” Journal of the Royal Statistical Society. Series B (Methodological), 102--107.

Lindley DV (1965). Introduction to probability and statistics: from a Bayesian viewpoint. 2. Inference. CUP Archive.

Author

Freddy Hernandez fhernanb@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rLIN(n=200, mu=2)

# Fitting the model
require(gamlss)
mod <- gamlss(y ~ 1, family="LIN")

#> GAMLSS-RS iteration 1: Global Deviance = 191.8689 

# Extracting the fitted values for mu
# using the inverse link function
exp(coef(mod, what='mu'))

#> (Intercept) 
#>    2.186758 

# Example 2
# Generating random values under some model
n <- 100
x1 <- runif(n=n)
x2 <- runif(n=n)
eta <- 1 + 3 * x1 - 2 * x2
mu <- exp(eta)
y <- rLIN(n=n, mu=mu)

mod <- gamlss(y ~ x1 + x2, family=LIN)

#> GAMLSS-RS iteration 1: Global Deviance = -69.9185 
#> GAMLSS-RS iteration 2: Global Deviance = -69.9185 

coef(mod, what='mu')

#> (Intercept)          x1          x2 
#>    1.142223    2.650584   -2.086987