The Marshall-Olkin Extended Inverse Weibull family

The Marshall-Olkin Extended Inverse Weibull family

Usage

MOEIW(mu.link = "log", sigma.link = "log", nu.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

nu.link

defines the nu.link, with "log" link as the default for the nu parameter.

Value

Returns a gamlss.family object which can be used to fit a MOEIW distribution in the gamlss() function.

Details

The Marshall-Olkin Extended Inverse Weibull distribution with parameters mu, sigma and nu has density given by

\(f(x) = \frac{\mu \sigma \nu x^{-(\sigma + 1)} exp\{{-\mu x^{-\sigma}}\}}{\{\nu -(\nu-1) exp\{{-\mu x ^{-\sigma}}\} \}^{2}},\)

for x > 0.

References

Hassan M O, A.H E, A.M.K T, Abdulkareem M Bc (2017). “Extended inverse Weibull distribution with reliability application.” Journal of the Egyptian Mathematical Society, 25, 343--349. doi:10.1016/j.joems.2017.02.006 , http://dx.doi.org/10.1016/j.joems.2017.02.006.

See also

Author

Amylkar Urrea Montoya, amylkar.urrea@udea.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rMOEIW(n=400, mu=0.6, sigma=1.7, nu=0.3)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family='MOEIW',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma and nu
# using the inverse link function
exp(coef(mod, what='mu'))
#> (Intercept) 
#>   0.7289874 
exp(coef(mod, what='sigma'))
#> (Intercept) 
#>    1.556561 
exp(coef(mod, what='nu'))
#> (Intercept) 
#>   0.2272875 

# Example 2
# Generating random values under some model
n <- 400
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(-2.02 + 3 * x1)
sigma <- exp(2.23 - 2 * x2)
nu <- 0.3
x <- rMOEIW(n=n, mu, sigma, nu)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, family=MOEIW,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
#> (Intercept)          x1 
#>   -2.321493    3.487804 
coef(mod, what="sigma")
#> (Intercept)          x2 
#>    2.107156   -1.648467 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>   0.3316438