Plot of initValOW objects

Draws the empirical total time on test (TTT) plot and its non-parametric (LOESS) estimated curve useful for searching initial values for OW family.

# S3 method for initValOW
plot(
  x,
  xlab = "i/n",
  ylab = expression(phi(u)),
  xlim = c(0, 1),
  ylim = c(0, 1),
  col = 1,
  lty = NULL,
  lwd = NA,
  main = "",
  curve_options = list(col = 2, lwd = 2, lty = 1),
  legend_options = list(pos = "top"),
  ...
)

Arguments

x
an object of class initVal, generated with initValuesOW_TTT.

xlab, ylab
titles for x and y axes, as in plot.

xlim
the x limits (x1, x2) of the plot.

ylim
the y limits (x1, x2) of the plot.

col
The colors for lines and points. Multiple colors can be specified. This is the usual color argument of plot.default.

lty
a vector of line types, see par for further information.

lwd
a vector of line widths, see par for further information.

main
a main title for the plot,

curve_options
a list with further arguments useful for customization of non-parametric estimate plot.

legend_options
a list with fur further arguments useful for customization of the legend of the plot.

...
further arguments passed empirical TTT plot.

Details

This plot complements the use of initValuesOW_TTT. It is always advisable to use this function in order to check the result of non-parametric estimate of TTT plot. See the first example in Examples section for an illustration.

The possible arguments for ... can be consulted in plot.default and par.