Summary of initValOW objects

This summary method displays initial values and search regions for OW family.

Usage

# S3 method for class 'initValOW'
summary(object, ...)

Arguments

object

an object of class initVal, generated with initValuesOW.

...

extra arguments

Value

No return value, it just prints out in the console the initial values and the search regions for \(sigma\) and \(nu\) from OW distribution (see dOW).

Author

Jaime Mosquera Gutiérrez jmosquerag@unal.edu.co