The Exponentiated Weibull family

The Exponentiated Weibull family

EW(mu.link = "log", sigma.link = "log", nu.link = "log")

Arguments

mu.link
defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link
defines the sigma.link, with "log" link as the default for the sigma.

nu.link
defines the nu.link, with "log" link as the default for the nu parameter.

Details

The Exponentiated Weibull Distribution with parameters mu, sigma and nu has density given by

\(f(x)=\nu \mu \sigma x^{\sigma-1} \exp(-\mu x^\sigma) (1-\exp(-\mu x^\sigma))^{\nu-1},\)

for x > 0.

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rEW(n=100, mu=2, sigma=1.5, nu=0.5)

# Fitting the model
mod <- gamlss::gamlss(y~1, sigma.fo=~1, nu.fo=~1, family='EW')

#> GAMLSS-RS iteration 1: Global Deviance = 31.6371 
#> GAMLSS-RS iteration 2: Global Deviance = 31.1579 
#> GAMLSS-RS iteration 3: Global Deviance = 30.8201 
#> GAMLSS-RS iteration 4: Global Deviance = 30.5379 
#> GAMLSS-RS iteration 5: Global Deviance = 30.2924 
#> GAMLSS-RS iteration 6: Global Deviance = 30.0695 
#> GAMLSS-RS iteration 7: Global Deviance = 29.8612 
#> GAMLSS-RS iteration 8: Global Deviance = 29.6771 
#> GAMLSS-RS iteration 9: Global Deviance = 29.51 
#> GAMLSS-RS iteration 10: Global Deviance = 29.3617 
#> GAMLSS-RS iteration 11: Global Deviance = 29.2337 
#> GAMLSS-RS iteration 12: Global Deviance = 29.113 
#> GAMLSS-RS iteration 13: Global Deviance = 29.0057 
#> GAMLSS-RS iteration 14: Global Deviance = 28.912 
#> GAMLSS-RS iteration 15: Global Deviance = 28.8276 
#> GAMLSS-RS iteration 16: Global Deviance = 28.7539 
#> GAMLSS-RS iteration 17: Global Deviance = 28.689 
#> GAMLSS-RS iteration 18: Global Deviance = 28.6315 
#> GAMLSS-RS iteration 19: Global Deviance = 28.5806 
#> GAMLSS-RS iteration 20: Global Deviance = 28.5351 

#> Warning: Algorithm RS has not yet converged

# Extracting the fitted values for mu, sigma and nu
# using the inverse link function
exp(coef(mod, what='mu'))

#> (Intercept) 
#>    1.977254 

exp(coef(mod, what='sigma'))

#> (Intercept) 
#>    1.551491 

exp(coef(mod, what='nu'))

#> (Intercept) 
#>   0.5720148