The Ex-Wald family

The function ExWALD() defines the Ex-wALD distribution, three-parameter continuous distribution for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

Usage

ExWALD(mu.link = "log", sigma.link = "log", nu.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma parameter.

nu.link

defines the nu.link, with "log" link as the default for the nu parameter.

Value

Returns a gamlss.family object which can be used to fit a Ex-WALD distribution in the gamlss() function.

Details

The Ex-Wald distribution with parameters \(\mu\), \(\sigma\) and \(\nu\) has density given by

\(f(x |\mu, \sigma) = x + 1 \)

References

Heathcote, A. (2004). Fitting Wald and ex-Wald distributions to response time data: An example using functions for the S-PLUS package. Behavior Research Methods, Instruments, & Computers, 36, 678-694.

See also

Author

Freddy Hernandez, fhernanb@unal.edu.co

Examples

# Example 1
# Generating random values with
# known mu, sigma and nu

mu <- 0.20
sigma <- 70
nu <- 115

y <- rExWALD(n=100, mu, sigma, nu)

library(gamlss)
mod1 <- gamlss(y~1, family=ExWALD,
               control=gamlss.control(n.cyc=1000, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 1284.621 

exp(coef(mod1, what="mu"))
#> (Intercept) 
#>   0.2048618 
exp(coef(mod1, what="sigma"))
#> (Intercept) 
#>    72.03127 
exp(coef(mod1, what="nu"))
#> (Intercept) 
#>    140.3192 

# Example 2
# Generating random values under some model

# \donttest{
# A function to simulate a data set with Y ~ ExWALD
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu    <- exp(-1 + 2.8 * x1) # 1.5 approximately
  sigma <- exp( 1 - 1.2 * x2) # 1.5 approximately
  nu    <- 2
  y <- rExWALD(n=n, mu=mu, sigma=sigma, nu=nu)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=100)

# Fitting the model
mod2 <- gamlss(y ~ x1,
               sigma.fo = ~ x2,
               nu.fo = ~ 1,
               family = ExWALD,
               data = dat,
               control = gamlss.control(n.cyc=1000, 
                                        trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 615.001 
#> GAMLSS-RS iteration 2: Global Deviance = 427.7696 
#> GAMLSS-RS iteration 3: Global Deviance = 406.2887 
#> GAMLSS-RS iteration 4: Global Deviance = 405.5854 
#> GAMLSS-RS iteration 5: Global Deviance = 405.2429 
#> GAMLSS-RS iteration 6: Global Deviance = 405.0361 
#> GAMLSS-RS iteration 7: Global Deviance = 404.9048 
#> GAMLSS-RS iteration 8: Global Deviance = 404.823 
#> GAMLSS-RS iteration 9: Global Deviance = 404.7733 
#> GAMLSS-RS iteration 10: Global Deviance = 404.7417 
#> GAMLSS-RS iteration 11: Global Deviance = 404.7216 
#> GAMLSS-RS iteration 12: Global Deviance = 404.7085 
#> GAMLSS-RS iteration 13: Global Deviance = 404.7002 
#> GAMLSS-RS iteration 14: Global Deviance = 404.6949 
#> GAMLSS-RS iteration 15: Global Deviance = 404.6914 
#> GAMLSS-RS iteration 16: Global Deviance = 404.6893 
#> GAMLSS-RS iteration 17: Global Deviance = 404.6876 
#> GAMLSS-RS iteration 18: Global Deviance = 404.6867 
summary(mod2)
#> Warning: summary: vcov has failed, option qr is used instead
#> ******************************************************************
#> Family:  c("ExWALD", "Ex-Wald") 
#> 
#> Call:  gamlss(formula = y ~ x1, sigma.formula = ~x2, nu.formula = ~1,  
#>     family = ExWALD, data = dat, control = gamlss.control(n.cyc = 1000,  
#>         trace = TRUE)) 
#> 
#> Fitting method: RS() 
#> 
#> ------------------------------------------------------------------
#> Mu link function:  log
#> Mu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  -0.8144     0.3179  -2.562   0.0119 *  
#> x1            3.0736     0.6930   4.435 2.41e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> Sigma link function:  log
#> Sigma Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   1.5540     0.2340   6.640 1.76e-09 ***
#> x2           -1.8191     0.5254  -3.463 0.000794 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> Nu link function:  log 
#> Nu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   0.7790     0.1122   6.941  4.1e-10 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> No. of observations in the fit:  100 
#> Degrees of Freedom for the fit:  5
#>       Residual Deg. of Freedom:  95 
#>                       at cycle:  18 
#>  
#> Global Deviance:     404.6867 
#>             AIC:     414.6867 
#>             SBC:     427.7126 
#> ******************************************************************
# }