The Kumaraswamy Inverse Weibull family

The Kumaraswamy Inverse Weibull family

Usage

KumIW(mu.link = "log", sigma.link = "log", nu.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

nu.link

defines the nu.link, with "log" link as the default for the nu parameter.

Value

Returns a gamlss.family object which can be used to fit a KumIW distribution in the gamlss() function.

Details

The Kumaraswamy Inverse Weibull Distribution with parameters mu, sigma and nu has density given by

\(f(x)= \mu \sigma \nu x^{-\sigma - 1} \exp{- \mu x^{-\sigma}} (1 - \exp{- \mu x^{-\sigma}})^{\nu - 1},\)

for \(x > 0\), \(\mu > 0\), \(\sigma > 0\) and \(\nu > 0\).

The KumIW distribution with \(\nu=1\) corresponds with the IW distribution.

References

almalki2014modificationsRelDists

shahbaz2012kumaraswamyRelDists

See also

Author

Freddy Hernandez, fhernanb@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rKumIW(n=1000, mu=1.5, sigma=0.3, nu=1)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family=KumIW,
              control=gamlss.control(n.cyc=5000, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 13136.25 
#> GAMLSS-RS iteration 2: Global Deviance = 12758.93 
#> GAMLSS-RS iteration 3: Global Deviance = 12756.74 
#> GAMLSS-RS iteration 4: Global Deviance = 12755.28 
#> GAMLSS-RS iteration 5: Global Deviance = 12753.99 
#> GAMLSS-RS iteration 6: Global Deviance = 12752.83 
#> GAMLSS-RS iteration 7: Global Deviance = 12751.79 
#> GAMLSS-RS iteration 8: Global Deviance = 12750.85 
#> GAMLSS-RS iteration 9: Global Deviance = 12750.02 
#> GAMLSS-RS iteration 10: Global Deviance = 12749.27 
#> GAMLSS-RS iteration 11: Global Deviance = 12748.6 
#> GAMLSS-RS iteration 12: Global Deviance = 12748 
#> GAMLSS-RS iteration 13: Global Deviance = 12747.47 
#> GAMLSS-RS iteration 14: Global Deviance = 12746.98 
#> GAMLSS-RS iteration 15: Global Deviance = 12746.55 
#> GAMLSS-RS iteration 16: Global Deviance = 12746.16 
#> GAMLSS-RS iteration 17: Global Deviance = 12745.81 
#> GAMLSS-RS iteration 18: Global Deviance = 12745.5 
#> GAMLSS-RS iteration 19: Global Deviance = 12745.22 
#> GAMLSS-RS iteration 20: Global Deviance = 12744.97 
#> GAMLSS-RS iteration 21: Global Deviance = 12744.74 
#> GAMLSS-RS iteration 22: Global Deviance = 12744.54 
#> GAMLSS-RS iteration 23: Global Deviance = 12744.35 
#> GAMLSS-RS iteration 24: Global Deviance = 12744.19 
#> GAMLSS-RS iteration 25: Global Deviance = 12744.04 
#> GAMLSS-RS iteration 26: Global Deviance = 12743.91 
#> GAMLSS-RS iteration 27: Global Deviance = 12743.79 
#> GAMLSS-RS iteration 28: Global Deviance = 12743.68 
#> GAMLSS-RS iteration 29: Global Deviance = 12743.58 
#> GAMLSS-RS iteration 30: Global Deviance = 12743.49 
#> GAMLSS-RS iteration 31: Global Deviance = 12743.41 
#> GAMLSS-RS iteration 32: Global Deviance = 12743.34 
#> GAMLSS-RS iteration 33: Global Deviance = 12743.27 
#> GAMLSS-RS iteration 34: Global Deviance = 12743.22 
#> GAMLSS-RS iteration 35: Global Deviance = 12743.16 
#> GAMLSS-RS iteration 36: Global Deviance = 12743.12 
#> GAMLSS-RS iteration 37: Global Deviance = 12743.07 
#> GAMLSS-RS iteration 38: Global Deviance = 12743.03 
#> GAMLSS-RS iteration 39: Global Deviance = 12743 
#> GAMLSS-RS iteration 40: Global Deviance = 12742.97 
#> GAMLSS-RS iteration 41: Global Deviance = 12742.94 
#> GAMLSS-RS iteration 42: Global Deviance = 12742.91 
#> GAMLSS-RS iteration 43: Global Deviance = 12742.89 
#> GAMLSS-RS iteration 44: Global Deviance = 12742.87 
#> GAMLSS-RS iteration 45: Global Deviance = 12742.85 
#> GAMLSS-RS iteration 46: Global Deviance = 12742.83 
#> GAMLSS-RS iteration 47: Global Deviance = 12742.81 
#> GAMLSS-RS iteration 48: Global Deviance = 12742.8 
#> GAMLSS-RS iteration 49: Global Deviance = 12742.79 
#> GAMLSS-RS iteration 50: Global Deviance = 12742.78 
#> GAMLSS-RS iteration 51: Global Deviance = 12742.77 
#> GAMLSS-RS iteration 52: Global Deviance = 12742.76 
#> GAMLSS-RS iteration 53: Global Deviance = 12742.75 
#> GAMLSS-RS iteration 54: Global Deviance = 12742.74 
#> GAMLSS-RS iteration 55: Global Deviance = 12742.73 
#> GAMLSS-RS iteration 56: Global Deviance = 12742.73 
#> GAMLSS-RS iteration 57: Global Deviance = 12742.72 
#> GAMLSS-RS iteration 58: Global Deviance = 12742.72 
#> GAMLSS-RS iteration 59: Global Deviance = 12742.71 
#> GAMLSS-RS iteration 60: Global Deviance = 12742.71 
#> GAMLSS-RS iteration 61: Global Deviance = 12742.7 
#> GAMLSS-RS iteration 62: Global Deviance = 12742.7 
#> GAMLSS-RS iteration 63: Global Deviance = 12742.7 
#> GAMLSS-RS iteration 64: Global Deviance = 12742.69 
#> GAMLSS-RS iteration 65: Global Deviance = 12742.69 
#> GAMLSS-RS iteration 66: Global Deviance = 12742.69 
#> GAMLSS-RS iteration 67: Global Deviance = 12742.69 
#> GAMLSS-RS iteration 68: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 69: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 70: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 71: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 72: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 73: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 74: Global Deviance = 12742.68 
#> GAMLSS-RS iteration 75: Global Deviance = 12742.67 

# Extracting the fitted values for mu, sigma and nu 
# using the inverse link function
exp(coef(mod, what="mu"))
#> (Intercept) 
#>    1.587114 
exp(coef(mod, what="sigma"))
#> (Intercept) 
#>   0.2893338 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>    1.007727 

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n)
x2 <- runif(n)
mu <- exp(1 +  -1 * x1)
sigma <- exp(1 + -1 * x2)
nu <- 5
y <- rKumIW(n=n, mu=mu, sigma=sigma, nu=nu)

mod <- gamlss(y~x1, sigma.fo=~x2, nu.fo=~1, family=KumIW,
              control=gamlss.control(n.cyc=5000, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 106.6089 
#> GAMLSS-RS iteration 2: Global Deviance = 86.9101 
#> GAMLSS-RS iteration 3: Global Deviance = 86.7592 
#> GAMLSS-RS iteration 4: Global Deviance = 86.6677 
#> GAMLSS-RS iteration 5: Global Deviance = 86.579 
#> GAMLSS-RS iteration 6: Global Deviance = 86.4931 
#> GAMLSS-RS iteration 7: Global Deviance = 86.4099 
#> GAMLSS-RS iteration 8: Global Deviance = 86.3294 
#> GAMLSS-RS iteration 9: Global Deviance = 86.2514 
#> GAMLSS-RS iteration 10: Global Deviance = 86.1759 
#> GAMLSS-RS iteration 11: Global Deviance = 86.1027 
#> GAMLSS-RS iteration 12: Global Deviance = 86.0318 
#> GAMLSS-RS iteration 13: Global Deviance = 85.9632 
#> GAMLSS-RS iteration 14: Global Deviance = 85.8966 
#> GAMLSS-RS iteration 15: Global Deviance = 85.8321 
#> GAMLSS-RS iteration 16: Global Deviance = 85.7696 
#> GAMLSS-RS iteration 17: Global Deviance = 85.709 
#> GAMLSS-RS iteration 18: Global Deviance = 85.6503 
#> GAMLSS-RS iteration 19: Global Deviance = 85.5933 
#> GAMLSS-RS iteration 20: Global Deviance = 85.5381 
#> GAMLSS-RS iteration 21: Global Deviance = 85.4845 
#> GAMLSS-RS iteration 22: Global Deviance = 85.4326 
#> GAMLSS-RS iteration 23: Global Deviance = 85.3822 
#> GAMLSS-RS iteration 24: Global Deviance = 85.3356 
#> GAMLSS-RS iteration 25: Global Deviance = 85.2903 
#> GAMLSS-RS iteration 26: Global Deviance = 85.2462 
#> GAMLSS-RS iteration 27: Global Deviance = 85.2034 
#> GAMLSS-RS iteration 28: Global Deviance = 85.1617 
#> GAMLSS-RS iteration 29: Global Deviance = 85.1211 
#> GAMLSS-RS iteration 30: Global Deviance = 85.0817 
#> GAMLSS-RS iteration 31: Global Deviance = 85.0433 
#> GAMLSS-RS iteration 32: Global Deviance = 85.0059 
#> GAMLSS-RS iteration 33: Global Deviance = 84.9696 
#> GAMLSS-RS iteration 34: Global Deviance = 84.9342 
#> GAMLSS-RS iteration 35: Global Deviance = 84.8998 
#> GAMLSS-RS iteration 36: Global Deviance = 84.8663 
#> GAMLSS-RS iteration 37: Global Deviance = 84.8337 
#> GAMLSS-RS iteration 38: Global Deviance = 84.802 
#> GAMLSS-RS iteration 39: Global Deviance = 84.771 
#> GAMLSS-RS iteration 40: Global Deviance = 84.7409 
#> GAMLSS-RS iteration 41: Global Deviance = 84.7116 
#> GAMLSS-RS iteration 42: Global Deviance = 84.6831 
#> GAMLSS-RS iteration 43: Global Deviance = 84.6553 
#> GAMLSS-RS iteration 44: Global Deviance = 84.6282 
#> GAMLSS-RS iteration 45: Global Deviance = 84.6019 
#> GAMLSS-RS iteration 46: Global Deviance = 84.5762 
#> GAMLSS-RS iteration 47: Global Deviance = 84.5512 
#> GAMLSS-RS iteration 48: Global Deviance = 84.5268 
#> GAMLSS-RS iteration 49: Global Deviance = 84.5031 
#> GAMLSS-RS iteration 50: Global Deviance = 84.4799 
#> GAMLSS-RS iteration 51: Global Deviance = 84.4574 
#> GAMLSS-RS iteration 52: Global Deviance = 84.4354 
#> GAMLSS-RS iteration 53: Global Deviance = 84.414 
#> GAMLSS-RS iteration 54: Global Deviance = 84.3931 
#> GAMLSS-RS iteration 55: Global Deviance = 84.3727 
#> GAMLSS-RS iteration 56: Global Deviance = 84.3529 
#> GAMLSS-RS iteration 57: Global Deviance = 84.3335 
#> GAMLSS-RS iteration 58: Global Deviance = 84.3147 
#> GAMLSS-RS iteration 59: Global Deviance = 84.2963 
#> GAMLSS-RS iteration 60: Global Deviance = 84.2783 
#> GAMLSS-RS iteration 61: Global Deviance = 84.2608 
#> GAMLSS-RS iteration 62: Global Deviance = 84.2438 
#> GAMLSS-RS iteration 63: Global Deviance = 84.2271 
#> GAMLSS-RS iteration 64: Global Deviance = 84.2109 
#> GAMLSS-RS iteration 65: Global Deviance = 84.1951 
#> GAMLSS-RS iteration 66: Global Deviance = 84.1796 
#> GAMLSS-RS iteration 67: Global Deviance = 84.1645 
#> GAMLSS-RS iteration 68: Global Deviance = 84.1498 
#> GAMLSS-RS iteration 69: Global Deviance = 84.1355 
#> GAMLSS-RS iteration 70: Global Deviance = 84.1215 
#> GAMLSS-RS iteration 71: Global Deviance = 84.1078 
#> GAMLSS-RS iteration 72: Global Deviance = 84.0945 
#> GAMLSS-RS iteration 73: Global Deviance = 84.0815 
#> GAMLSS-RS iteration 74: Global Deviance = 84.0687 
#> GAMLSS-RS iteration 75: Global Deviance = 84.0563 
#> GAMLSS-RS iteration 76: Global Deviance = 84.0442 
#> GAMLSS-RS iteration 77: Global Deviance = 84.0324 
#> GAMLSS-RS iteration 78: Global Deviance = 84.0209 
#> GAMLSS-RS iteration 79: Global Deviance = 84.0096 
#> GAMLSS-RS iteration 80: Global Deviance = 83.9986 
#> GAMLSS-RS iteration 81: Global Deviance = 83.9878 
#> GAMLSS-RS iteration 82: Global Deviance = 83.9773 
#> GAMLSS-RS iteration 83: Global Deviance = 83.9671 
#> GAMLSS-RS iteration 84: Global Deviance = 83.957 
#> GAMLSS-RS iteration 85: Global Deviance = 83.9473 
#> GAMLSS-RS iteration 86: Global Deviance = 83.9377 
#> GAMLSS-RS iteration 87: Global Deviance = 83.9284 
#> GAMLSS-RS iteration 88: Global Deviance = 83.9192 
#> GAMLSS-RS iteration 89: Global Deviance = 83.9103 
#> GAMLSS-RS iteration 90: Global Deviance = 83.9016 
#> GAMLSS-RS iteration 91: Global Deviance = 83.8931 
#> GAMLSS-RS iteration 92: Global Deviance = 83.8848 
#> GAMLSS-RS iteration 93: Global Deviance = 83.8767 
#> GAMLSS-RS iteration 94: Global Deviance = 83.8687 
#> GAMLSS-RS iteration 95: Global Deviance = 83.861 
#> GAMLSS-RS iteration 96: Global Deviance = 83.8534 
#> GAMLSS-RS iteration 97: Global Deviance = 83.846 
#> GAMLSS-RS iteration 98: Global Deviance = 83.8387 
#> GAMLSS-RS iteration 99: Global Deviance = 83.8316 
#> GAMLSS-RS iteration 100: Global Deviance = 83.8247 
#> GAMLSS-RS iteration 101: Global Deviance = 83.8179 
#> GAMLSS-RS iteration 102: Global Deviance = 83.8113 
#> GAMLSS-RS iteration 103: Global Deviance = 83.8048 
#> GAMLSS-RS iteration 104: Global Deviance = 83.7985 
#> GAMLSS-RS iteration 105: Global Deviance = 83.7923 
#> GAMLSS-RS iteration 106: Global Deviance = 83.7862 
#> GAMLSS-RS iteration 107: Global Deviance = 83.7803 
#> GAMLSS-RS iteration 108: Global Deviance = 83.7745 
#> GAMLSS-RS iteration 109: Global Deviance = 83.7688 
#> GAMLSS-RS iteration 110: Global Deviance = 83.7625 
#> GAMLSS-RS iteration 111: Global Deviance = 83.7566 
#> GAMLSS-RS iteration 112: Global Deviance = 83.7507 
#> GAMLSS-RS iteration 113: Global Deviance = 83.7451 
#> GAMLSS-RS iteration 114: Global Deviance = 83.7395 
#> GAMLSS-RS iteration 115: Global Deviance = 83.7341 
#> GAMLSS-RS iteration 116: Global Deviance = 83.7288 
#> GAMLSS-RS iteration 117: Global Deviance = 83.7237 
#> GAMLSS-RS iteration 118: Global Deviance = 83.7186 
#> GAMLSS-RS iteration 119: Global Deviance = 83.7137 
#> GAMLSS-RS iteration 120: Global Deviance = 83.7089 
#> GAMLSS-RS iteration 121: Global Deviance = 83.7042 
#> GAMLSS-RS iteration 122: Global Deviance = 83.6996 
#> GAMLSS-RS iteration 123: Global Deviance = 83.6951 
#> GAMLSS-RS iteration 124: Global Deviance = 83.6908 
#> GAMLSS-RS iteration 125: Global Deviance = 83.6865 
#> GAMLSS-RS iteration 126: Global Deviance = 83.6823 
#> GAMLSS-RS iteration 127: Global Deviance = 83.6783 
#> GAMLSS-RS iteration 128: Global Deviance = 83.6743 
#> GAMLSS-RS iteration 129: Global Deviance = 83.6704 
#> GAMLSS-RS iteration 130: Global Deviance = 83.6666 
#> GAMLSS-RS iteration 131: Global Deviance = 83.6629 
#> GAMLSS-RS iteration 132: Global Deviance = 83.6592 
#> GAMLSS-RS iteration 133: Global Deviance = 83.6557 
#> GAMLSS-RS iteration 134: Global Deviance = 83.6522 
#> GAMLSS-RS iteration 135: Global Deviance = 83.6488 
#> GAMLSS-RS iteration 136: Global Deviance = 83.6455 
#> GAMLSS-RS iteration 137: Global Deviance = 83.6423 
#> GAMLSS-RS iteration 138: Global Deviance = 83.6391 
#> GAMLSS-RS iteration 139: Global Deviance = 83.6361 
#> GAMLSS-RS iteration 140: Global Deviance = 83.633 
#> GAMLSS-RS iteration 141: Global Deviance = 83.6301 
#> GAMLSS-RS iteration 142: Global Deviance = 83.6272 
#> GAMLSS-RS iteration 143: Global Deviance = 83.6244 
#> GAMLSS-RS iteration 144: Global Deviance = 83.6216 
#> GAMLSS-RS iteration 145: Global Deviance = 83.619 
#> GAMLSS-RS iteration 146: Global Deviance = 83.6163 
#> GAMLSS-RS iteration 147: Global Deviance = 83.6138 
#> GAMLSS-RS iteration 148: Global Deviance = 83.6113 
#> GAMLSS-RS iteration 149: Global Deviance = 83.6085 
#> GAMLSS-RS iteration 150: Global Deviance = 83.6058 
#> GAMLSS-RS iteration 151: Global Deviance = 83.6032 
#> GAMLSS-RS iteration 152: Global Deviance = 83.6007 
#> GAMLSS-RS iteration 153: Global Deviance = 83.5982 
#> GAMLSS-RS iteration 154: Global Deviance = 83.5958 
#> GAMLSS-RS iteration 155: Global Deviance = 83.5934 
#> GAMLSS-RS iteration 156: Global Deviance = 83.5911 
#> GAMLSS-RS iteration 157: Global Deviance = 83.5888 
#> GAMLSS-RS iteration 158: Global Deviance = 83.5867 
#> GAMLSS-RS iteration 159: Global Deviance = 83.5845 
#> GAMLSS-RS iteration 160: Global Deviance = 83.5824 
#> GAMLSS-RS iteration 161: Global Deviance = 83.5804 
#> GAMLSS-RS iteration 162: Global Deviance = 83.5784 
#> GAMLSS-RS iteration 163: Global Deviance = 83.5765 
#> GAMLSS-RS iteration 164: Global Deviance = 83.5746 
#> GAMLSS-RS iteration 165: Global Deviance = 83.5728 
#> GAMLSS-RS iteration 166: Global Deviance = 83.571 
#> GAMLSS-RS iteration 167: Global Deviance = 83.5693 
#> GAMLSS-RS iteration 168: Global Deviance = 83.5677 
#> GAMLSS-RS iteration 169: Global Deviance = 83.5661 
#> GAMLSS-RS iteration 170: Global Deviance = 83.5645 
#> GAMLSS-RS iteration 171: Global Deviance = 83.5629 
#> GAMLSS-RS iteration 172: Global Deviance = 83.5614 
#> GAMLSS-RS iteration 173: Global Deviance = 83.56 
#> GAMLSS-RS iteration 174: Global Deviance = 83.5585 
#> GAMLSS-RS iteration 175: Global Deviance = 83.5571 
#> GAMLSS-RS iteration 176: Global Deviance = 83.5558 
#> GAMLSS-RS iteration 177: Global Deviance = 83.5544 
#> GAMLSS-RS iteration 178: Global Deviance = 83.5531 
#> GAMLSS-RS iteration 179: Global Deviance = 83.5518 
#> GAMLSS-RS iteration 180: Global Deviance = 83.5506 
#> GAMLSS-RS iteration 181: Global Deviance = 83.5494 
#> GAMLSS-RS iteration 182: Global Deviance = 83.5482 
#> GAMLSS-RS iteration 183: Global Deviance = 83.547 
#> GAMLSS-RS iteration 184: Global Deviance = 83.5459 
#> GAMLSS-RS iteration 185: Global Deviance = 83.5448 
#> GAMLSS-RS iteration 186: Global Deviance = 83.5437 
#> GAMLSS-RS iteration 187: Global Deviance = 83.5427 
#> GAMLSS-RS iteration 188: Global Deviance = 83.5416 
#> GAMLSS-RS iteration 189: Global Deviance = 83.5406 
#> GAMLSS-RS iteration 190: Global Deviance = 83.5396 

coef(mod, what="mu")
#> (Intercept)          x1 
#>    1.001282   -1.074256 
coef(mod, what="sigma")
#> (Intercept)          x2 
#>     1.03179    -1.08353 
exp(coef(mod, what="nu"))
#> (Intercept) 
#>    4.682122