Custimized region search for odd Weibull distribution

This function can be used to modify OW gamlss.family object in order to set a customized region search for gamlss() function.

myOW_region(family = OW, valid.values = "auto", initVal)

Arguments

family

The OW family. This arguments allows the user to modify input arguments of the family, like the link functions.

valid.values

a list of character elements specifying the region for sigma and/or nu. See Details and Examples section to learn about its use.

initVal

An initValOW object generated with initValuesOW_TTT function.

Details

This function was created to help users to fit OW distribution easily bounding the parametric space for sigma and nu.

The valid.values must be defined as a list of characters containing a call of the all function.

Author

Jaime Mosquera Gutiérrez jmosquerag@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rOW(n=200, mu=0.2, sigma=4, nu=0.05)

# Custom search region
myvalues <- list(sigma="all(sigma > 1)",
                 nu="all(nu < 1) & all(nu < 1)")

my_initial_guess <- initValuesOW_TTT(formula=y~1)
summary(my_initial_guess)
#> --------------------------------------------------------------------
#> Initial Values
#> sigma = 5
#> nu = 0.1
#> --------------------------------------------------------------------
#> Search Regions
#> For sigma: all(sigma > 1)
#> For nu: all(nu < 1/sigma)
#> --------------------------------------------------------------------
#> Hazard shape: Bathtub

# OW family modified with 'myOW_region'
require(gamlss)
myOW <- myOW_region(valid.values=myvalues, initVal=my_initial_guess)
mod1 <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, 
               sigma.start=param.startOW('sigma', my_initial_guess), 
               nu.start=param.startOW('nu', my_initial_guess),
               control=gamlss.control(n.cyc=300, trace=FALSE),
               family=myOW)
#> Error in while (abs(G.dev.old - G.dev) > c.crit && iter < n.cyc) {    if ("mu" %in% names(family$parameters)) {        if (family$parameter$mu == TRUE & mu.fix == FALSE) {            mu.fit <<- glim.fit(f = mu.object, X = mu.X, y = y,                 w = w, fv = mu, os = mu.offset, step = mu.step,                 control = i.control, gd.tol = gd.tol, auto = autostep)            mu <<- mu.fit$fv            mu.object$smooth <- mu.fit$smooth        }    }    if ("sigma" %in% names(family$parameters)) {        if (family$parameter$sigma == TRUE & sigma.fix == FALSE) {            sigma.fit <<- glim.fit(f = sigma.object, X = sigma.X,                 y = y, w = w, fv = sigma, os = sigma.offset,                 step = sigma.step, control = i.control, gd.tol = gd.tol,                 auto = autostep)            sigma <<- sigma.fit$fv            sigma.object$smooth <- sigma.fit$smooth        }    }    if ("nu" %in% names(family$parameters)) {        if (family$parameter$nu == TRUE & nu.fix == FALSE) {            nu.fit <<- glim.fit(f = nu.object, X = nu.X, y = y,                 w = w, fv = nu, os = nu.offset, step = nu.step,                 control = i.control, gd.tol = gd.tol, auto = autostep)            nu <<- nu.fit$fv            nu.object$smooth <- nu.fit$smooth        }    }    if ("tau" %in% names(family$parameters)) {        if (family$parameter$tau == TRUE & tau.fix == FALSE) {            tau.fit <<- glim.fit(f = tau.object, X = tau.X, y = y,                 w = w, fv = tau, os = tau.offset, step = tau.step,                 control = i.control, gd.tol = gd.tol, auto = autostep)            tau <<- tau.fit$fv            tau.object$smooth <- tau.fit$smooth        }    }    G.dev.old <- G.dev    G.dev.incr <- eval(G.dev.expr)    G.dev <- sum(w * G.dev.incr)    iter <- iter + 1    fiter <<- iter    if (trace)         cat("GAMLSS-RS iteration ", iter, ": Global Deviance = ",             format(round(G.dev, 4)), " \n", sep = "")    if (G.dev > (G.dev.old + gd.tol) && iter > 1)         stop(paste("The global deviance is increasing", "\n",             "Try different steps for the parameters or the model maybe inappropriate"))}: missing value where TRUE/FALSE needed

exp(coef(mod1, what='mu'))
#> Error in coef(mod1, what = "mu"): object 'mod1' not found
exp(coef(mod1, what='sigma'))
#> Error in coef(mod1, what = "sigma"): object 'mod1' not found
exp(coef(mod1, what='nu'))
#> Error in coef(mod1, what = "nu"): object 'mod1' not found

# Example 2
# Same example using another link function and using 'myOW_region'
# in the argument 'family'
mod2 <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, 
               sigma.start=2, nu.start=0.1,
               control=gamlss.control(n.cyc=300, trace=FALSE),
               family=myOW_region(family=OW(sigma.link='identity'),
                                  valid.values=myvalues,
                                  initVal=my_initial_guess))

exp(coef(mod2, what='mu'))
#> (Intercept) 
#>   0.2146868 
coef(mod2, what='sigma')
#> (Intercept) 
#>    3.560331 
exp(coef(mod2, what='nu'))
#> (Intercept) 
#>   0.0580717