Custimized region search for odd Weibull distribution

This function can be used to modify OW gamlss.family object in order to set a customized region search for gamlss() function.

myOW_region(family = OW, valid.values = "auto", initVal)

Arguments

family
The OW family. This arguments allows the user to modify input arguments of the family, like the link functions.

valid.values
a list of character elements specifying the region for sigma and/or nu. See Details and Examples section to learn about its use.

initVal
an initValOW object generated with initValuesOW_TTT function.

Details

This function was created to help users to fit OW distribution easily bounding the parametric space for sigma and nu.

The valid.values must be defined as a list of characters containing a call of the all function.

Examples

# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rOW(n=200, mu=0.2, sigma=4, nu=0.05)

# Custom search region
myvalues <- list(sigma="all(sigma > 1)",
                 nu="all(nu < 1) & all(nu < 1)")

my_initial_guess <- initValuesOW_TTT(formula=y~1)

# OW family modified with 'myOW_region'
require(gamlss)
myOW <- myOW_region(valid.values=myvalues, initVal=my_initial_guess)
mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1,
              sigma.start=2, nu.start=0.1,
              control=gamlss.control(n.cyc=300, trace=FALSE),
              family=myOW)

exp(coef(mod, what='mu'))

#> (Intercept) 
#>   0.1955959 

exp(coef(mod, what='sigma'))

#> (Intercept) 
#>    4.327328 

exp(coef(mod, what='nu'))

#> (Intercept) 
#>  0.04508682 

# Example 2
# Same example using another link function and using 'myOW_region'
# in the argument 'family'
mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1,
              sigma.start=2, nu.start=0.1,
              control=gamlss.control(n.cyc=300, trace=FALSE),
              family=myOW_region(family=OW(sigma.link='identity'),
                                 valid.values=myvalues,
                                 initVal=my_initial_guess))

exp(coef(mod, what='mu'))

#> (Intercept) 
#>   0.1956048 

coef(mod, what='sigma')

#> (Intercept) 
#>    4.327067 

exp(coef(mod, what='nu'))

#> (Intercept) 
#>  0.04508918